% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-ss-box.R
\name{add_ss_box}
\alias{add_ss_box}
\alias{calc_joint_CI95}
\title{Add a Sensitivity/Specificity Box}
\usage{
add_ss_box(x, col = "black", alpha = 0.35)

calc_joint_CI95(sens, spec, n.controls, n.cases)
}
\arguments{
\item{x}{A \verb{2x2} data frame or \code{tibble} containing the lower
and upper CI95 joint confidence limits for sensitivity and specificity.
A call to \code{\link[=calc_joint_CI95]{calc_joint_CI95()}} generates values in this specified format.}

\item{col}{\code{character(1)} or \code{integer(1)}. Specify the colors for
lines, points, bar, box, or ROC.}

\item{alpha}{\code{numeric(1)} in \verb{[0, 1]}. The color transparency.
See also \code{\link[ggplot2:reexports]{ggplot2::alpha()}}.}

\item{sens}{\code{numeric(n)}. The sensitivity: \eqn{[0, 1]}.}

\item{spec}{\code{numeric(n)}. The specificity: \eqn{[0, 1]}.}

\item{n.controls}{\code{integer(1)}. Number of control or non-cases.}

\item{n.cases}{\code{integer(1)}. Number of cases/disease.}
}
\value{
A \eqn{2x2} matrix containing rows of sensitivity and
specificity respectively and columns of lower and upper 95\%
joint confidence intervals respectively.
}
\description{
Add a shaded box, typically to a ROC curve, that
corresponds to the 95\% joint binomial confidence interval
of the sensitivity and specificity.

Calculate the joint 95\% confidence interval
given sensitivity and specificity.
}
\details{
Recall that the ROC curve is \code{1 - specificity}, therefore the
added box involves internally inverting the specificity limits
so that the interval matches the plot.
}
\examples{
g <- ggplot2::ggplot(data.frame(x = 0.2, y = 0.8), ggplot2::aes(x = x, y = y)) +
  ggplot2::geom_point(shape = 18, size = 3) +
  ggplot2::lims(x = 0:1, y = 0:1) +
  ggplot2::labs(y = "Sensitivity", x = "1 - Specificity")
g

# calculate CI95s for 80/80 sens/spec
ci95 <- calc_joint_CI95(0.8, 0.8, 35, 65)
ci95

# unequal box due to class imbalance (65/35)
g + add_ss_box(ci95, col = "blue", alpha = 0.25)
}
\seealso{
\code{\link[=calc_ci_binom]{calc_ci_binom()}}
}
\author{
Stu Field, Amanda Hiser

Mike Mehan
}
