% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSensSpecBox.R
\name{addSensSpecBox}
\alias{addSensSpecBox}
\alias{calcJointCI95}
\title{Add a Sensitivity/Specificity Box}
\usage{
addSensSpecBox(x, col = "black", alpha = 0.35)

calcJointCI95(sens, spec, n.controls, n.cases)
}
\arguments{
\item{x}{A \verb{2x2} data frame containing the lower and upper CI95
joint confidence limits for sensitivity and specificity.
A call to \code{\link[=calcJointCI95]{calcJointCI95()}} generates this matrix in the specified format.}

\item{col}{Character or integer. Specify the colors for
lines, points, bar, box, or ROC.}

\item{alpha}{Numeric in \verb{[0, 1]}. The color transparency level.
See also \code{\link[ggplot2:reexports]{ggplot2::alpha()}}.}

\item{sens}{Numeric. The sensitivity: \verb{[0, 1]}.}

\item{spec}{Numeric. The specificity: \verb{[0, 1]}.}

\item{n.controls}{Integer. Number of control or non-cases.}

\item{n.cases}{Integer. Number of cases/disease.}
}
\value{
A \verb{2x2} matrix containing rows of sensitivity and
specificity respectively and columns of lower and upper 95\%
joint confidence intervals respectively.
}
\description{
Add a shaded box, typically to a ROC curve, that
corresponds to the 95\% joint binomial confidence interval
of the sensitivity and specificity.

Calculate the joint 95\% confidence interval
given sensitivity and specificity.
}
\details{
Recall that the ROC curve is \code{1 - specificity}, therefore the
added box involves internally inverting the specificity limits
so that the interval matches the plot.
}
\examples{
g <- ggplot2::ggplot(data.frame(x = 0.2, y = 0.8), ggplot2::aes(x = x, y = y)) +
  ggplot2::geom_point(shape = 18, size = 3) +
  ggplot2::lims(x = 0:1, y = 0:1) +
  ggplot2::labs(y = "Sensitivity", x = "1 - Specificity")
g

# calculate CI95s for 80/80 sens/spec
ci95 <- calcJointCI95(0.8, 0.8, 35, 65)
ci95

# unequal box due to class imbalance (65/35)
g + addSensSpecBox(ci95, col = "blue", alpha = 0.25)
}
\seealso{
\code{\link[=calcBinomCI]{calcBinomCI()}}
}
\author{
Stu Field, Amanda Hiser

Mike Mehan
}
