% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotAUCs.R
\name{barplotAUCs}
\alias{barplotAUCs}
\title{Plot AUCs and Error Bars}
\usage{
barplotAUCs(
  data,
  color = SomaPlotr::soma_colors$lightgrey,
  flip = FALSE,
  main = bquote("AUCs ± CI95")
)
}
\arguments{
\item{data}{A \code{data.frame} object of AUCs and 95\% confidence intervals.
Each row is the result of a call to \code{\link[=calcEmpAUC]{calcEmpAUC()}} with
\code{ci95 = TRUE} and converted to a single row \code{data.frame}. See example.}

\item{color}{Character or numeric vector containing colors for each of
the barplots, as used by \code{\link[=ggplot]{ggplot()}}.
Vector length should match the number of rows in \code{data}.
Colors are recycled as necessary.}

\item{flip}{Logical. Should the axes be flipped? See example.}

\item{main}{Character. Optional string for the plot title.}
}
\value{
A \code{\link[=ggplot]{ggplot()}} plot.
}
\description{
Plots a list of AUCs as barplots and added error bars for each
corresponding to the 95\% confidence interval for each. See
\code{\link[=calcEmpAUC]{calcEmpAUC()}} for how to generate the AUCs and CI95s.
}
\examples{
# create random AUCs and CI95s
withr::with_seed(22, {
  true <- sample(c("control", "disease"), 20, replace = TRUE)
  auc_df <- lapply(1:5, function(.x) {
    data.frame(calcEmpAUC(true, runif(20), "disease", ci95 = TRUE))
  }) |> do.call(what = rbind)
})
auc_df
barplotAUCs(auc_df)

# Set rownames to identify the bars
rownames(auc_df) <- LETTERS[1:nrow(auc_df)]
barplotAUCs(auc_df)

# Flip axes
barplotAUCs(auc_df, color = SomaPlotr::soma_colors$purple, flip = TRUE)
}
\seealso{
\code{\link[=calcEmpAUC]{calcEmpAUC()}}
}
\author{
Stu Field
}
