% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-cor.R
\name{calc.cor}
\alias{calc.cor}
\alias{print.cor_table}
\alias{writeStatTable.cor_table}
\title{Calculate Correlation Table}
\usage{
calc.cor(
  data,
  apts = NULL,
  response,
  bh = TRUE,
  method = c("spearman", "pearson", "kendall"),
  ...
)

\method{print}{cor_table}(x, n = 6L, ...)

\method{writeStatTable}{cor_table}(x, file)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. String identifying the continuous response variable.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{method}{Matched String. The correlation method to use.
Matched to one of:
\itemize{
\item \verb{"spearman"}
\item \verb{"pearson"}
\item \verb{"kendall"}
}}

\item{...}{Additional arguments passed to \code{\link[=cor.test]{cor.test()}}.}

\item{x}{An object of class \code{cor_table}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below)}
\item{test}{The name of the statistical test performed, e.g. "KS-test".}
\item{call}{The direct call made the \verb{calc.*()}.}
\item{paired}{Logical. Whether the values are paired (if applicable).}
\item{response}{The "Response" column used in the statistical test.}
\item{counts}{Group counts split by the "Response" column (if applicable).}
\item{log}{Logical. Whether the RFU values were log10-transformed.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data frame used for testing.}

The \code{stat.table} contains:
\itemize{
\item \strong{Test Statistic}: KS-distance, t-statistic, U, W, H, r, rho, tau, etc.
\item \strong{Signed Statistic}: Signed test statistic (if applicable).
\item \strong{p.value}: Unadjusted p-values for the test statistic.
\item \strong{fdr or q.value}: FDR adjustment to p-values (either BH or Storey).
\item \strong{p.bonferroni}: Bonferroni corrected p-value.
\item \strong{rank}: Ranking by p-value.
\item \strong{\code{loCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
\item \strong{\code{hiCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
}
}
\description{
Generate a correlation table, representing the relationship between a
continuous dependent variable, and each of the aptamers in the menu
in the ADAT (columns).
}
\section{Functions}{
\itemize{
\item \code{print(cor_table)}: S3 print method for Spearman, Pearson, and Kendall correlation
tables, all \code{cor_table} objects.

\item \code{writeStatTable(cor_table)}: The S3 \code{writeStatTable} method for class \code{cor_table}.

}}
\note{
The dependent variable (\code{response}) is \emph{not} log-transformed prior
to analysis. If this is desired, the user is directed to do so in
the data object itself prior to the call.
}
\examples{
cor_table <- list()

# Spearman
cor_table$spearman <- withr::with_options(
  c(warn = -1),
  calc.cor(sim_test_data, response = "reg_response")
)

# Pearson
cor_table$pearson  <- calc.cor(log10(sim_test_data),
                               response = "reg_response", method = "pearson")

# S3 print method
cor_table   # both spearman and pearson

}
\seealso{
\code{\link[=cor.test]{cor.test()}}

The `calc.*()` family: 
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field
}
\concept{calc}
