% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-glm.R
\name{calc.glm}
\alias{calc.glm}
\alias{print.glm_table}
\alias{write_stat_table.glm_table}
\alias{plot.glm_table}
\title{Calculate Logistic Regression Table}
\usage{
calc.glm(data, response = NULL, apts = NULL, bh = TRUE)

\method{print}{glm_table}(x, n = 6L, ...)

\method{write_stat_table}{glm_table}(x, file)

\method{plot}{glm_table}(
  x,
  .data,
  n_plots = seq(6),
  pt.size = 2,
  pt.pch = 19,
  fit.col = SomaPlotr::soma_colors$magenta,
  ...
)
}
\arguments{
\item{data}{The data frame (or \code{soma_adat}) containing data to be plotted.
In many cases this may now be included in the \code{lm_table} object itself.}

\item{response}{Character. String identifying the categorical response variable.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{x}{An object of class \code{glm_table} created via call \code{\link[=calc.glm]{calc.glm()}}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{...}{Arguments required by the S3 generics.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}

\item{.data}{The data frame containing data to be plotted.
In many cases this may now be included in the \code{"glm.table"} object itself.}

\item{n_plots}{Numeric. Numeric indices corresponding to the \code{glm_table}
object, "stat.table" data frame rows. For example, \code{1:6} will plot the top
6 analytes from the table.}

\item{pt.size}{Numeric. Size for the points.}

\item{pt.pch}{Shape of points. \code{\link[=par]{par()}}.}

\item{fit.col}{Character or numeric. Color of the model fit line,
as used by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below).}
\item{models}{The stripped down (for size) model objects for each analyte.}
\item{call}{The direct call made to \verb{calc.*()}.}
\item{test}{The name of the statistical test; linear regression.}
\item{y.response}{The column name of the dependent/response variable.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data matrix.}
\item{log}{Whether the RFU values were log10-transformed prior to analysis.}

The \code{stat.table} contains:
+ \strong{intercept}: the \eqn{\beta_0} intercept of the model.
+ \strong{slope}: the \eqn{\beta_1} slope of the model.
+ \strong{odds_ratio}: the odds ratio for classes (\code{\link[=calc.glm]{calc.glm()}} only).
+ \strong{p.value}: Unadjusted p-values for the test statistic
corresponding to \eqn{\beta_1}.
+ \strong{fdr or q.value}: FDR adjusted p-value (Benjamini-Hochberg) or
if q-value, Storey.
+ \strong{p.bonferroni}: Bonferroni corrected p-value.
+ \strong{rank}: Ranking by the test significance.
}
\description{
Generate a table of logistic regression models
representing the relationship between a two-class
response variable, and each of the \code{predictors} in the data (columns).
}
\section{Functions}{
\itemize{
\item \code{print(glm_table)}: S3 print method for \code{glm_table} objects.

\item \code{write_stat_table(glm_table)}: The S3 \code{write_stat_table} method for class \code{glm_table}.

\item \code{plot(glm_table)}: S3 plotting method for objects of class \code{glm_table}.

}}
\examples{
# convert to training data object for log-reg downstream & log-transform
tr <- create_train(log10(sim_test_data), group.var = gender)
glm_table <- calc.glm(tr)

# S3 print method
glm_table                 # top 6

# extend the default no. of rows
print(glm_table, n = 10)  # top 10

# S3 plot method
plot(glm_table, .data = tr, n_plots = c(2, 4, 6))
}
\seealso{
\code{\link[=fit_logistic]{fit_logistic()}}, \code{\link[=glm]{glm()}}

The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field
}
\concept{calc}
