% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-kw.R
\name{calc.kw}
\alias{calc.kw}
\alias{print.kw_table}
\alias{writeStatTable.kw_table}
\title{Create Kruskal-Wallis Table}
\usage{
calc.kw(data, apts = NULL, response = NULL, bh = TRUE, ...)

\method{print}{kw_table}(x, n = 6L, ...)

\method{writeStatTable}{kw_table}(x, file)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. The column name to use as the response
grouping. Expected to be a factor with > 2 levels.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{...}{Additional arguments passed to \code{\link[=kruskal.test]{kruskal.test()}}.}

\item{x}{An object of class \code{kw_table}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below)}
\item{test}{The name of the statistical test performed, e.g. "KS-test".}
\item{call}{The direct call made the \verb{calc.*()}.}
\item{paired}{Logical. Whether the values are paired (if applicable).}
\item{response}{The "Response" column used in the statistical test.}
\item{counts}{Group counts split by the "Response" column (if applicable).}
\item{log}{Logical. Whether the RFU values were log10-transformed.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data frame used for testing.}

The \code{stat.table} contains:
\itemize{
\item \strong{Test Statistic}: KS-distance, t-statistic, U, W, H, r, rho, tau, etc.
\item \strong{Signed Statistic}: Signed test statistic (if applicable).
\item \strong{p.value}: Unadjusted p-values for the test statistic.
\item \strong{fdr or q.value}: FDR adjustment to p-values (either BH or Storey).
\item \strong{p.bonferroni}: Bonferroni corrected p-value.
\item \strong{rank}: Ranking by p-value.
\item \strong{\code{loCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
\item \strong{\code{hiCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
}
}
\description{
Computes the Kruskal-Wallis univariate test statistic
given a response column for each variable in the data matrix.
One-way ANOVA on ranks (non-parametric).
}
\section{Functions}{
\itemize{
\item \code{print(kw_table)}: S3 print method for \code{kw_table} objects.

\item \code{writeStatTable(kw_table)}: The S3 \code{writeStatTable} method for class \code{kw_table}.

}}
\examples{
# Create multi-level variable
sim_test_data$Response <- factor(paste(sim_test_data$class_response,
                                       sim_test_data$gender, sep = "_"))
print(levels(sim_test_data$Response))

kw_table <- calc.kw(sim_test_data, response = "Response")

# S3 print method
kw_table

}
\references{
Myles Hollander and Douglas A. Wolfe (1973),
\emph{Nonparametric Statistical Methods}. New York: John Wiley & Sons. Pages 115-120.
}
\seealso{
\code{\link[=kruskal.test]{kruskal.test()}}

The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field
}
\concept{calc}
