% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-lm.R
\name{calc.lm}
\alias{calc.lm}
\alias{print.lm_table}
\alias{writeStatTable.lm_table}
\alias{plot.lm_table}
\title{Calculate Linear Models Table}
\usage{
calc.lm(data, apts = NULL, response, bh = TRUE)

\method{print}{lm_table}(x, n = 6, ...)

\method{writeStatTable}{lm_table}(x, file)

\method{plot}{lm_table}(
  x,
  .data,
  n_plots = seq(6L),
  method = "loess",
  y.lab = "RFU",
  fit.col = SomaPlotr::soma_colors$purple,
  ci.col = SomaPlotr::soma_colors$lightgrey,
  ci = TRUE,
  pt.pch = 19,
  pt.col = "black",
  pt.size = 2,
  ...
)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. String identifying the \emph{continuous}, dependent
response variable. The \code{response} is \emph{not} log-transformed prior
to analysis. If this is desired, the user is directed to do so in
the data object itself prior to the call.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{x}{An object of class \code{lm_table} created via call to \code{\link[=calc.lm]{calc.lm()}}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{...}{Arguments required by the S3 generics.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}

\item{.data}{The data frame (or \code{soma_adat}) containing data to be plotted.
In many cases this may now be included in the \code{lm_table} object itself.}

\item{n_plots}{Numeric. Numeric indices corresponding
to the \code{lm_table} object, "stat.table" data frame rows.
For example, \code{1:6}, will plot the top 6 analytes from the table.}

\item{method}{Character. See \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}.}

\item{y.lab}{Character. An optional string for the y-axis label.}

\item{fit.col}{Character or numeric. Color of the model fit line,
as used by \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}

\item{ci.col}{Character. Color of the interval line to be plotted,
as used in \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{ci}{Logical. Should standard error based confidence
intervals be plotted? See \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}.}

\item{pt.pch}{Integer. Shape for of points.}

\item{pt.col}{Character. Color for the points.}

\item{pt.size}{Numeric. Size for the points.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below).}
\item{models}{The stripped down (for size) model objects for each analyte.}
\item{call}{The direct call made to \verb{calc.*()}.}
\item{test}{The name of the statistical test; linear regression.}
\item{y.response}{The column name of the dependent/response variable.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data matrix.}
\item{log}{Whether the RFU values were log10-transformed prior to analysis.}

The \code{stat.table} contains:
+ \strong{intercept}: the \eqn{\beta_0} intercept of the model.
+ \strong{slope}: the \eqn{\beta_1} slope of the model.
+ \strong{odds_ratio}: the odds ratio for classes (\code{\link[=calc.glm]{calc.glm()}} only).
+ \strong{p.value}: Unadjusted p-values for the test statistic
corresponding to \eqn{\beta_1}.
+ \strong{fdr or q.value}: FDR adjusted p-value (Benjamini-Hochberg) or
if q-value, Storey.
+ \strong{p.bonferroni}: Bonferroni corrected p-value.
+ \strong{rank}: Ranking by the test significance.
}
\description{
Generate a table of linear regression models of the classic form:
\deqn{Y_i ~ \beta_0 + \beta_1 * X_i + \epsilon_i}
where in this context,
\deqn{Y = response}
\deqn{X = predictor}
This model represents the relationship between a continuous dependent
variable (\code{response}), and each of the analyte features (\code{predictor}) in
the ADAT. The "response" in this case is typically some clinical or assay
variable that can be predicted by an analyte RFU
via \eqn{\beta_0} and \eqn{\beta_1}.
}
\section{Functions}{
\itemize{
\item \code{print(lm_table)}: S3 print method for \code{lm_table} objects.

\item \code{writeStatTable(lm_table)}: The S3 \code{writeStatTable} method for class \code{lm_table}.

\item \code{plot(lm_table)}: S3 plotting method for objects of class \code{lm_table}.

}}
\examples{
lm_table <- calc.lm(log10(sim_test_data), response = "HybControlNormScale")

# S3 print method
lm_table

# extend the default no. of rows
print(lm_table, n = 10)

# S3 writeStatTable method
apt_data <- getAnalyteInfo(sim_test_data)
lm_table$stat.table <- addTargetInfo(lm_table$stat.table, apt_data)
f_out <- tempfile("lm-table-", fileext = ".csv")
writeStatTable(lm_table, file = f_out)

# S3 plot method
plot(lm_table, .data = sim_test_data)

# Adjust the `method` argument to an `lm` fit as in the model
plot(lm_table, .data = sim_test_data, method = "lm")

# There are many options to the defaults
plot(lm_table, .data = sim_test_data, n_plots = c(2, 5, 9), pt.col = "darkred",
     ci.col = "green", pt.size = 4, fit.col = "cyan", pt.pch = 13,
     y.lab = "Changed y-axis"
)
}
\seealso{
\code{\link[=lm]{lm()}}

The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field
}
\concept{calc}
