% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-lr.R
\name{calc.lr}
\alias{calc.lr}
\alias{print.logratio_table}
\alias{writeStatTable.logratio_table}
\title{Calculate log-ratio Table}
\usage{
calc.lr(
  data,
  apts = NULL,
  response = NULL,
  paired = FALSE,
  do.mean = FALSE,
  ...
)

\method{print}{logratio_table}(x, n = 6, ...)

\method{writeStatTable}{logratio_table}(x, file)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. The column name to use as the
response grouping. Expected to be a factor with 2 levels.
If not a factor, \code{response} will be coerced to a
factor (i.e. alphabetical order). Level 2 of the factor is
considered the \code{disease/case} group, thus positive values
for the statistic indicate up-regulation in \code{Level2} compared \code{Level1}.}

\item{paired}{Logical. Indicating whether the data are paired or not.}

\item{do.mean}{Logical. Should the mean be used (rather than median) for
calculating log-ratios?}

\item{...}{Additional arguments passed to the internal \code{column_lr},
typically \code{rm.outliers = TRUE} if desired.}

\item{x}{An object of class \code{logratio_table}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
An object of class \code{c(stat_table, logratio_table)} sorted by the
log-fold changes for the response variable.
}
\description{
Create a statistical table of the log fold changes for a data set
given the response column with 2 factor levels.
}
\section{Functions}{
\itemize{
\item \code{print(logratio_table)}: S3 print method for \code{logratio_table} objects.

\item \code{writeStatTable(logratio_table)}: The S3 \code{writeStatTable} method for class \code{logratio_table}.

}}
\examples{
lr1 <- calc.lr(sim_test_data, response = "class_response")

# remove outliers
lr2 <- calc.lr(sim_test_data, response = "class_response", rm.outliers = TRUE)

# S3 print method
lr1
lr2

# S3 writeStatTable method
apt_data <- getAnalyteInfo(sim_test_data)
lr1$stat.table <- addTargetInfo(lr1$stat.table, apt_data)
f_out <- tempfile("logRatio-table-", fileext = ".csv")
writeStatTable(lr1, file = f_out)
}
\seealso{
The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field
}
\concept{calc}
