% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mackwolfe.R
\name{calc.mackwolfe}
\alias{calc.mackwolfe}
\alias{calc.jt}
\alias{print.mackwolfe_table}
\alias{write_stat_table.mackwolfe_table}
\title{Create Mack-Wolfe Table}
\usage{
calc.mackwolfe(data, apts = NULL, response = NULL, peak, bh = TRUE, ...)

calc.jt(data, apts = NULL, response = "SampleGroup", bh = TRUE, ...)

\method{print}{mackwolfe_table}(x, n = 6L, ...)

\method{write_stat_table}{mackwolfe_table}(x, file)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. The name of the column in \code{data}
containing the grouping information for the Mack-Wolfe test.}

\item{peak}{Character. A string corresponding to the desired peak factor
level (\emph{not} an integer, a string). If \code{peak = "jt"} a
JT-test is performed. If \code{peak = NULL}, the "peak unknown" version of
the Mack-Wolfe test is performed.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{...}{Additional arguments passed to \code{\link[=mackwolfe]{mackwolfe()}}, particularly
\code{verbose} or \code{nperm}.}

\item{x}{An object of class \code{mackwolfe_table}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
A list object of class \code{c(stat_table, mackwolfe_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results.}
\item{call}{The direct call made the \code{calc.mackwolfe}.}
\item{test}{The name of the test performed, "Mack-Wolfe Test".}
\item{data.dim}{The dimensions of the data frame used for testing.}
\item{peak}{The peak factor level assumed by the test.}
\item{response}{The "Response" column used in the statistical test.}
\item{factor.order}{The ordering of the "Response" column factor levels.}
\item{counts}{Group counts split by the "Response" column.}
\item{data.frame}{The name of the data frame containing the RFU data.}

The \code{stat.table} contains:
+ \strong{Ap}: The Mack-Wolfe test statistics.
+ \strong{Astar}: The z-transform of \code{Ap}.
+ \strong{n}: Number of samples used in calculating \code{Ap}.
+ \strong{peak}: The string level of the peak and its factor level (\code{k}).
+ \strong{p.value}: Unadjusted p-values for the \code{Astar} test statistics.
+ \strong{fdr or q.value}: FDR adjustment to p-values (either BH or Storey).
+ \strong{p.bonferroni}: Bonferroni corrected p-value.
+ \strong{rank}: Ranking by the test significance.
}
\description{
Computes the Mack-Wolfe univariate non-parametric test statistic
given a response column and ordering for each variable in the data matrix.
}
\details{
Distribution free test of a peak or trends (JT-test). If the JT-test
is desired, set \code{peak = "jt"}.
}
\section{Functions}{
\itemize{
\item \code{print(mackwolfe_table)}: S3 print method for \code{mackwolfe_table} objects.

\item \code{write_stat_table(mackwolfe_table)}: The S3 \code{write_stat_table} method for class \code{mackwolfe_table}.

}}
\examples{
lett <- head(LETTERS, 4)
sim_test_data$group <- factor(rep(lett, each = nrow(sim_test_data) / 4L))

# Peak at "B"
mack <- calc.mackwolfe(sim_test_data, response = "group", peak = "B")

# S3 print method
mack

# Permutation P-value estimation
# This is time consuming to run for all features
# so we use only the top features from above
# with nperm > 1000 is recommended
feats  <- rownames(head(mack$stat.table, 10L))
new_df <- dplyr::select(sim_test_data, group, all_of(feats))
mack_perm <- calc.mackwolfe(new_df, response = "group", nperm = 100, peak = "B")

# Thin wrapper for JT-test
# peak = "jt"
calc.jt(sim_test_data, response = "group")

}
\references{
Myles Hollander and Douglas A. Wolfe (1973).
\emph{Nonparametric Statistical Methods}. New York: John Wiley & Sons. Pages 115-120.
}
\seealso{
\code{\link[=mackwolfe]{mackwolfe()}}

The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field
}
\concept{calc}
