% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-robust-lm.R
\name{calc.robust.lm}
\alias{calc.robust.lm}
\alias{print.rlm_table}
\alias{writeStatTable.rlm_table}
\title{Calculate Robust Linear Models Table}
\usage{
calc.robust.lm(data, apts = NULL, response, bh = TRUE, do.log = deprecated())

\method{print}{rlm_table}(x, n = 6, ...)

\method{writeStatTable}{rlm_table}(x, file)
}
\arguments{
\item{data}{The data frame (or \code{soma_adat}) containing data to be plotted.
In many cases this may now be included in the \code{lm_table} object itself.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. String identifying the \emph{continuous}, dependent
response variable. The \code{response} is \emph{not} log-transformed prior
to analysis. If this is desired, the user is directed to do so in
the data object itself prior to the call.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{do.log}{Deprecated.}

\item{x}{An object of class \code{rlm_table} created via call to \code{\link[=calc.robust.lm]{calc.robust.lm()}}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{...}{Arguments required by the S3 generics.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below).}
\item{models}{The stripped down (for size) model objects for each analyte.}
\item{call}{The direct call made to \verb{calc.*()}.}
\item{test}{The name of the statistical test; linear regression.}
\item{y.response}{The column name of the dependent/response variable.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data matrix.}
\item{log}{Whether the RFU values were log10-transformed prior to analysis.}

The \code{stat.table} contains:
+ \strong{intercept}: the \eqn{\beta_0} intercept of the model.
+ \strong{slope}: the \eqn{\beta_1} slope of the model.
+ \strong{odds_ratio}: the odds ratio for classes (\code{\link[=calc.glm]{calc.glm()}} only).
+ \strong{p.value}: Unadjusted p-values for the test statistic
corresponding to \eqn{\beta_1}.
+ \strong{fdr or q.value}: FDR adjusted p-value (Benjamini-Hochberg) or
if q-value, Storey.
+ \strong{p.bonferroni}: Bonferroni corrected p-value.
+ \strong{rank}: Ranking by the test significance.
}
\description{
Generate a table of linear regression models of the classic form:
\deqn{Y_i ~ \beta_0 + \beta_1 * X_i + \epsilon_i}
where in this context,
\deqn{Y = response}
\deqn{X = predictor}
This model represents the relationship between a continuous dependent
variable (\code{response}), and each of the analyte features (\code{predictor}) in
the ADAT. The "response" in this case is typically some clinical or assay
variable that can be predicted by an analyte RFU
via \eqn{\beta_0} and \eqn{\beta_1}.
}
\details{
Robust regression utilizes \code{\link[=rlm]{rlm()}}, where
"fitting is done by iterated re-weighted least-squares".
}
\section{Functions}{
\itemize{
\item \code{print(rlm_table)}: S3 print method for \code{rlm_table} objects.

\item \code{writeStatTable(rlm_table)}: The S3 \code{writeStatTable} method for class \code{rlm_table}.

}}
\examples{
rlm_table <- calc.robust.lm(log10(sim_test_data), response = "HybControlNormScale")

# S3 print method
rlm_table                 # top 6
print(rlm_table, n = 20)  # top 20

# S3 writeStatTable method
apt_data <- getAnalyteInfo(sim_test_data)
rlm_table$stat.table <- addTargetInfo(rlm_table$stat.table, apt_data)
f_out <- tempfile("rlm-table-", fileext = ".csv")
writeStatTable(rlm_table, file = f_out)
}
\seealso{
\code{\link[=lm]{lm()}}

The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.t}()},
\code{\link{calc.wilcox}()}
}
\author{
Stu Field, Leigh Alexander
}
\concept{calc}
