% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-t.R
\name{calc.t}
\alias{calc.t}
\alias{print.t_table}
\alias{writeStatTable.t_table}
\title{Calculate Student's t-statistics Table}
\usage{
calc.t(
  data,
  apts = NULL,
  bh = TRUE,
  paired = FALSE,
  response = "Response",
  do.log = deprecated(),
  ...
)

\method{print}{t_table}(x, n = 6, ...)

\method{writeStatTable}{t_table}(x, file)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{paired}{Logical. Indicating whether the data are paired or not.}

\item{response}{Character. The column name to use as the
response grouping. Expected to be a factor with 2 levels.
If not a factor, \code{response} will be coerced to a
factor (i.e. alphabetical order). Level 2 of the factor is
considered the \code{disease/case} group, thus positive values
for the statistic indicate up-regulation in \code{Level2} compared \code{Level1}.}

\item{do.log}{Deprecated.}

\item{...}{The \code{rm.outliers = TRUE} option should be passed here, otherwise
additional arguments to allow extensibility to S3 methods.}

\item{x}{An object of class \code{t_table}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below)}
\item{test}{The name of the statistical test performed, e.g. "KS-test".}
\item{call}{The direct call made the \verb{calc.*()}.}
\item{paired}{Logical. Whether the values are paired (if applicable).}
\item{response}{The "Response" column used in the statistical test.}
\item{counts}{Group counts split by the "Response" column (if applicable).}
\item{log}{Logical. Whether the RFU values were log10-transformed.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data frame used for testing.}

The \code{stat.table} contains:
\itemize{
\item \strong{Test Statistic}: KS-distance, t-statistic, U, W, H, r, rho, tau, etc.
\item \strong{Signed Statistic}: Signed test statistic (if applicable).
\item \strong{p.value}: Unadjusted p-values for the test statistic.
\item \strong{fdr or q.value}: FDR adjustment to p-values (either BH or Storey).
\item \strong{p.bonferroni}: Bonferroni corrected p-value.
\item \strong{rank}: Ranking by p-value.
\item \strong{\code{loCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
\item \strong{\code{hiCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
}
}
\description{
Computes t-statistics for a supplied data set (ADAT).
}
\section{Functions}{
\itemize{
\item \code{print(t_table)}: S3 print method for \code{t_table} objects.

\item \code{writeStatTable(t_table)}: The S3 \code{writeStatTable} method for class \code{t_table}.

}}
\examples{
t_table <- calc.t(log10(sim_test_data), response = "class_response")

# minor differences; some outliers removed
t2 <- calc.t(log10(sim_test_data), response = "class_response", rm.outliers = TRUE)
# S3 print method
t_table
t2

# S3 writeStatTable method
apt_data <- getAnalyteInfo(sim_test_data)
t_table$stat.table <- addTargetInfo(t_table$stat.table, apt_data)
f_out <- tempfile("t-table-", fileext = ".csv")
writeStatTable(t_table, file = f_out)
}
\seealso{
The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.wilcox}()}
}
\author{
Mike Mehan and Stu Field
}
\concept{calc}
