% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-wilcox.R
\name{calc.wilcox}
\alias{calc.wilcox}
\alias{calc.mw}
\alias{print.wilcox_table}
\alias{write_stat_table.wilcox_table}
\title{Calculate Wilcoxon/Mann-Whitney Table}
\usage{
calc.wilcox(data, apts = NULL, paired = FALSE, response = NULL, bh = TRUE, ...)

calc.mw(...)

\method{print}{wilcox_table}(x, n = 6L, ...)

\method{write_stat_table}{wilcox_table}(x, file)
}
\arguments{
\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{paired}{Logical. Indicating whether the data are paired or not.}

\item{response}{Character. The column name to use as the
response grouping. Expected to be a factor with 2 levels.
If not a factor, \code{response} will be coerced to a
factor (i.e. alphabetical order). Level 2 of the factor is
considered the \code{disease/case} group, thus positive values
for the statistic indicate up-regulation in \code{Level2} compared \code{Level1}.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{...}{Additional arguments passed to ultimately \code{\link[=wilcox.test]{wilcox.test()}}, or
the \code{rm.outliers = TRUE} option can also be passed. Otherwise arguments
required by other S3 generics.}

\item{x}{An object of class \code{wilcox_table}.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below)}
\item{test}{The name of the statistical test performed, e.g. "KS-test".}
\item{call}{The direct call made the \verb{calc.*()}.}
\item{paired}{Logical. Whether the values are paired (if applicable).}
\item{response}{The "Response" column used in the statistical test.}
\item{counts}{Group counts split by the "Response" column (if applicable).}
\item{log}{Logical. Whether the RFU values were log10-transformed.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data frame used for testing.}

The \code{stat.table} contains:
\itemize{
\item \strong{Test Statistic}: KS-distance, t-statistic, U, W, H, r, rho, tau, etc.
\item \strong{Signed Statistic}: Signed test statistic (if applicable).
\item \strong{p.value}: Unadjusted p-values for the test statistic.
\item \strong{fdr or q.value}: FDR adjustment to p-values (either BH or Storey).
\item \strong{p.bonferroni}: Bonferroni corrected p-value.
\item \strong{rank}: Ranking by p-value.
\item \strong{\code{loCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
\item \strong{\code{hiCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
}
}
\description{
Computes either the Wilcoxon signed rank test (if \code{paired = TRUE})
or the Wilcoxon rank sum test (if \code{paired = FALSE}; default)
aka the Mann-Whitney U. The function \code{\link[=calc.mw]{calc.mw()}} is a wrapper
for \code{\link[=calc.wilcox]{calc.wilcox()}} with the \code{paired = FALSE} (default) and is
maintained for backward compatibility.

A wrapper for calc.wilcox() and backward compatibility.
}
\details{
The following is from \code{\link[=wilcox.test]{wilcox.test()}}:\cr

If only \code{x} is given, or if both x and y are given and
\code{paired = TRUE}, a Wilcoxon signed rank test of the null that
the distribution of x (in the one sample case) or of \code{x - y} (in
the paired two sample case) is symmetric about mu is performed.

Otherwise, if both x and y are given and \code{paired = FALSE},
a Wilcoxon rank sum test (equivalent to the Mann-Whitney test)
is carried out. In this case, the null hypothesis is
that the distributions of x and y differ by a location shift
of mu and the alternative is that they differ by some other
location shift (and the one-sided alternative "greater" is that
x is shifted to the right of y).

By default (if exact is not specified), an exact p-value is
computed if the samples contain less than 50 finite values and
there are no ties. Otherwise, a normal approximation is used.
}
\section{Functions}{
\itemize{
\item \code{print(wilcox_table)}: S3 print method for \code{wilcox_table} objects.

\item \code{write_stat_table(wilcox_table)}: The S3 \code{write_stat_table} method for class \code{wilcox_table}.

}}
\examples{
# Suppress p-value ties warning
w  <- calc.wilcox(sim_test_data, response = "class_response")

# Mann-Whitney wrapper
mw <- calc.mw(sim_test_data, response = "class_response")  # same

# Remove outliers
w2 <- calc.wilcox(sim_test_data, response = "class_response", rm.outliers = TRUE)

# S3 print method
w
mw
w2

}
\seealso{
\code{\link[=wilcox.test]{wilcox.test()}}

The `calc.*()` family: 
\code{\link{calc.cor}()},
\code{\link{calc.glm}()},
\code{\link{calc.ks}()},
\code{\link{calc.kw}()},
\code{\link{calc.lm}()},
\code{\link{calc.lr}()},
\code{\link{calc.mackwolfe}()},
\code{\link{calc.robust.lm}()},
\code{\link{calc.t}()}
}
\author{
Stu Field
}
\concept{calc}
