% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-auc.R
\name{calc_auc}
\alias{calc_auc}
\alias{calc_emp_auc}
\alias{calc_pepe_auc}
\alias{calc_boot_auc}
\title{Calculate Area Under Curve}
\usage{
calc_auc(truth, predicted)

calc_emp_auc(truth, predicted, pos.class, ci95 = FALSE)

calc_pepe_auc(truth, predicted, pos.class)

calc_boot_auc(
  truth,
  predicted,
  pos.class,
  nboot = 1000,
  r.seed = sample(1000, 1)
)
}
\arguments{
\item{truth}{\code{character(n)} or \code{factor(n)}. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{\code{numeric(n)}. A numeric vector of class probabilities.}

\item{pos.class}{\code{character(1)}. Name of the "positive" or "event" class.}

\item{ci95}{\code{logical(1)}. Should DeLong's standard error based confidence
limits be included with the AUC estimate?}

\item{nboot}{\code{integer(1)}. The number of bootstrap estimates to perform.}
}
\value{
All return a numeric scalar corresponding to the area under
the curve. For 95\% confidence intervals (\code{ci95 = TRUE}), \code{\link[=calc_emp_auc]{calc_emp_auc()}}
returns a \code{list} object with these elements:
\item{auc}{The area under the curve (empirical).}
\item{lower.limit}{lower 95\% confidence limit based on standard error AUC.}
\item{upper.limit}{upper 95\% confidence limit based on standard error AUC.}

A list containing bootstrap intervals based on the number
of bootstraps performed:
\item{auc}{The raw Pepe AUC estimate from original data.}
\item{lower.limit}{The lower CI95 of the estimate.}
\item{upper.limit}{The upper CI95 of the estimate.}
}
\description{
Calculates the area under the curve (AUC).
}
\section{Functions}{
\itemize{
\item \code{calc_emp_auc()}: Calculate the \emph{empirical} AUC, optionally with corresponding 95\%
confidence intervals according to the DeLong approach via the standard
error of the AUC estimate. This empirical AUC estimate is calculated via
the trapezoid area at each step along the x-axis of a ROC curve.

\item \code{calc_pepe_auc()}: Calculate the AUC according to Margaret Pepe's book.

\item \code{calc_boot_auc()}: Bootstrapped confidence intervals for the 95\% limits are calculated via
\emph{empirical} bootstrap iterations and using Pepe's AUC calculation.

}}
\note{
\code{\link[=calc_auc]{calc_auc()}} is designed specifically, and only (!),
for binary 2-class problems.
}
\examples{
n <- 20
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
calc_auc(true, pred)

# Empirical AUC
calc_emp_auc(true, pred, "disease")
calc_emp_auc(true, pred, "disease", ci95 = TRUE)  # with CI95

# Pepe's AUC
calc_pepe_auc(true, pred, "disease")

# bootstrapped AUC
calc_boot_auc(true, pred, "disease")
calc_boot_auc(true, pred, "disease", nboot = 100, r.seed = 100)  # reproducible
}
\references{
DeLong et al. (1988) for the calculation of the Standard Error
of the Area Under the Curve (AUC) and of the difference between two AUCs.

\code{\link[=calc_pepe_auc]{calc_pepe_auc()}}: M. Pepe. The Statistical
Evaluation of Medical Tests for Classification and Prediction.
}
\seealso{
\code{\link[=plot_emp_roc]{plot_emp_roc()}}, \code{\link[=roc_xy]{roc_xy()}}

\code{\link[=replicate]{replicate()}}, \code{\link[=plot_emp_roc]{plot_emp_roc()}}
}
\author{
Stu Field
}
\concept{auc}
