% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-brier.R
\name{calc_brier}
\alias{calc_brier}
\title{Calculate the Brier Score}
\usage{
calc_brier(x, p)
}
\arguments{
\item{x}{\code{numeric(n)}. A vector of binary class data representing
the true classes. Must be all 0 or 1 or numeric coerceable.}

\item{p}{\code{numeric(n)}. A vector of the predicted probabilities,
i.e. in \verb{[0, 1]}.}
}
\value{
The Brier Score, a value in \verb{[0, 1]}, representing
the error in predictions, \code{0} being best possible score.
}
\description{
Calculates the Brier Score, i.e. MSE of probabilities in
\verb{[0, 1]} of a vector of probabilities.
The brier score attempts to account for not only wither
a prediction correctly predicts a class label at some
arbitrary evaluation cutoff, but also \emph{how close} the prediction
is to predicting the label, i.e. distinguishing between
\eqn{p = 0.51} and \eqn{p = 0.99}, despite both predicting
a positive class label.
}
\examples{
withr::with_seed(1, {
  n <- 100L
  p <- runif(n)
  x <- sample(0:1, n, replace = TRUE)
})
calc_brier(x, p)
}
\references{
https://en.wikipedia.org/wiki/Brier_score
}
\author{
Stu Field
}
