% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ci-binom.R
\name{calc_ci_binom}
\alias{calc_ci_binom}
\title{Calculate Binomial Confidence Interval}
\usage{
calc_ci_binom(p, n, ci = sqrt(0.95))
}
\arguments{
\item{p}{Numeric. The classification metric in \verb{[0, 1]}. Can also be a
vector of values representing the metric of interest (sens or spec).}

\item{n}{Integer. The total number of counts in the denominator
for the metric being calculated.}

\item{ci}{Numeric. The width of the confidence interval
to be calculated. Must be within \verb{[0.5, 1]}.}
}
\value{
A \code{tibble} object of the upper and lower binomial
confidence limits corresponding to the value of \code{ci}.
}
\description{
Calculates the \emph{joint} binomial confidence interval based
on the binomial variance given the data. Uses normal
approximation of the binomial.
}
\examples{
tp <- 16
fn <- 4
sens <- tp / (tp + fn)
calc_ci_binom(sens, tp + fn)
}
\references{
Margaret Pepe. 2003. The Statistical Evaluation of Medical
Tests for Classification and Prediction. (maybe?)
}
\seealso{
\code{\link[=qnorm]{qnorm()}}
}
\author{
Stu Field
}
