% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-confusion.R
\name{calc_confusion}
\alias{calc_confusion}
\alias{print.confusion_matrix}
\alias{summary.confusion_matrix}
\alias{print.summary_confusion_matrix}
\title{Calculate confusion matrix}
\usage{
calc_confusion(truth, predicted, pos.class, cutoff = 0.5)

\method{print}{confusion_matrix}(x, ...)

\method{summary}{confusion_matrix}(object, ...)

\method{print}{summary_confusion_matrix}(x, ...)
}
\arguments{
\item{truth}{\code{character(n)} or \code{factor(n)}. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{\code{numeric(n)}. A numeric vector of class probabilities.}

\item{pos.class}{\code{character(1)}. Name of the "positive" or "event" class.}

\item{cutoff}{\code{numeric(1)}. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{x}{A \code{confusion_matrix} or \code{summary_confusion_matrix} class object.}

\item{...}{Arguments passed to the \code{print} and \code{summary} generics.}

\item{object}{A \code{confusion_matrix} object, created via
\code{\link[=calc_confusion]{calc_confusion()}}.}
}
\value{
An object of class \code{confusion_matrix}, with the \emph{True} values
along the y-axis and \emph{Predicted} values along the x-axis.

Summary method returns a list object of class
\code{summary_confusion_matrix} consisting of:
\item{confusion:}{The class counts based on the confusion matrix.}
\item{metrics:}{Performance metric estimates, \code{n}, and associated
binomial 95\% confidence intervals. Note that \code{MCC} has a range in
\verb{[-1, 1]}, therefore confidence intervals are not calculated for this
metric (\code{\link[=calc_ci_binom]{calc_ci_binom()}} expects a probability value).}
\item{stats:}{F-measure, G-mean, and Weighted Accuracy.}
}
\description{
Calculate a confusion matrix from a series of predictions,
true class names and a decision cutoff. Returns a
\verb{2x2} confusion matrix. It is imperative that the
\emph{positive} class is clearly defined to avoid  ambiguity
with factor levels, therefore passing \verb{pos.class =} argument
is \emph{not} optional.
See \code{Details} for assumptions about the table layout.
}
\details{
Assume a \verb{2x2} table with notation:

\tabular{rcc}{
\tab Predicted \tab \cr
Truth    \tab negative  \tab positive \cr
negative \tab \emph{TN}      \tab \emph{FP} \cr
positive \tab \emph{FN}      \tab \emph{TP}
}

where:
\deqn{TN = True Negative}
\deqn{TP = True Positive}
\deqn{FN = False Negative}
\deqn{FP = False Positive}

The \code{summary} calculations are:
\deqn{Sensitivity = Recall = TP / (TP + FN)}
\deqn{Specificity = TN / (TN + FP)}
\deqn{Precision = PPV = TP / (TP + FP)}
\deqn{NPV = TN / (TN + FN)}
\deqn{Accuracy = (TP + TN) / (TP + FN + TN + FP)}
\deqn{Balanced Accuracy = (Sensitivity + Specificity) / 2}
\deqn{Prevalence = (FN + TP) / (TP + FN + TN + FP)}
\deqn{Matthew's Correlation Coefficient = TP x TN - FP x FN /
                            sqrt( (TP + FP) (TP + FN) (TN + FP) (TN + FN) )}
}
\section{Functions}{
\itemize{
\item \code{print(confusion_matrix)}: S3 print method for classes \code{confusion_matrix}.

\item \code{summary(confusion_matrix)}: Calculates the confusion statistics from a confusion matrix.

\item \code{print(summary_confusion_matrix)}: S3 print method for class \code{summary_confusion_matrix}.

}}
\examples{
n <- 20
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
(c_mat <- calc_confusion(true, pred, pos.class = "disease"))

calc_confusion(true, pred, pos.class = "disease", 0.75)    # specify cutoff

# factor levels of `truth` are ignored
# The `pos.class` argument is respected always
true_a <- factor(true, levels = c("control", "disease"))
true_b <- factor(true, levels = c("disease", "control"))
a <- calc_confusion(true_a, pred, pos.class = "disease")
b <- calc_confusion(true_b, pred, pos.class = "disease")
identical(a, b)

# S3 summary method
summary(c_mat)
}
\references{
The Statistical Evaluation of Medical Tests for Classification
and Prediction. 2004. Margaret Pepe, Altman, DG, Bland, JM. 1994.
"Diagnostic tests 1: sensitivity and specificity", British Medical
Journal, vol 308, 1552.
}
\seealso{
\code{\link[=calc_confusion]{calc_confusion()}}
}
\author{
Stu Field
}
