% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-univariate.R
\name{calc_univariate}
\alias{calc_univariate}
\title{Create Table of Univariate Results}
\usage{
calc_univariate(
  data,
  var,
  test = c("t.test", "lm", "ks", "kw", "logistic", "wilcox", "mw", "mackwolfe", "log2fc",
    "cor"),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} object containing data for analysis.}

\item{var}{\code{character(1)}. A response variable, a column in \code{data}.}

\item{test}{\code{character(1)}. A statistical test to run.
See above for currently supports tests.}

\item{...}{Additional parameters passed to the statistic
function defined in \code{test}.}
}
\value{
A \code{tibble} of features and univariate test results.
Common columns are:
\item{\code{p_value}}{The univariate p-value}
\item{\code{FDR}}{The false discovery rate corrected p-value}
\item{\code{p_bonferroni}}{The Bonferroni corrected p-value}
\item{\code{rank}}{The univariate rank of the feature}

Test-specific statistics include the following:
\item{\code{t.test}}{returns the t statistic, \code{t}}
\item{\code{lm}}{returns the intercept, slope, and t statistic of the slope,
\code{intercept}, \code{slope}, and \code{t_slope} respectively}
\item{\code{log2fc}}{returns the log2-fold-change of the ratio of group medians}
\item{\code{Kruskal-Wallis}}{returns the \code{H} test statistic}
\item{\code{KS}}{returns the \code{D} test statistic}
\item{\code{Wilcox}}{returns the \code{U} test statistic}
\item{\code{Mack-Wolfe}}{returns the \code{Ap} test statistic}
}
\description{
Iterates over features to compute univariate tests for each
given an appropriate response variable.
Currently supported tests:
\itemize{
\item{t-tests for binary endpoints}
\item{linear models for continuous endpoints}
\item{log2FC for the ratio of group medians)}
\item{Kruskal-Wallis for non-parametric multi-group comparisons}
\item{KS for for non-parametric binary comparisons}
\item{Wilcox for non-parametric binary comparisons (Mann-Whitney)}
\item{Mack-Wolfe for non-parametric trends (JT-test)}
}
}
\examples{
calc_univariate(mtcars, "vs")

calc_univariate(mtcars, "vs", "ks")

calc_univariate(mtcars, "mpg", "lm")

calc_univariate(mtcars, "cyl", "kw")

calc_univariate(mtcars, "vs", "wilcox")

calc_univariate(mtcars, "vs", "log2")

# grouping variable must be factor for mack-wolfe
mtcars$carb <- as.factor(mtcars$carb)
calc_univariate(mtcars, "carb", "mack", peak = "jt")

# for logistic: `var` is the LHS of the formula
calc_univariate(mtcars, "vs", "logistic")

calc_univariate(mtcars, "mpg", "cor")

# method = 'spearman' can be passed via '...'
calc_univariate(mtcars, "mpg", "cor", method = "spear")
}
\author{
Stu Field
}
