% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAUC.R
\name{calcAUC}
\alias{calcAUC}
\alias{calcEmpAUC}
\alias{calcPepeAUC}
\alias{calcBootAUC}
\title{Calculate Area Under Curve}
\usage{
calcAUC(truth, predicted, ...)

calcEmpAUC(truth, predicted, pos.class, ci95 = FALSE)

calcPepeAUC(truth, predicted, pos.class)

calcBootAUC(
  truth,
  predicted,
  pos.class,
  nboot = 1000,
  r.seed = sample(1000, 1)
)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{...}{Additional arguments passed to \code{\link[caTools:colAUC]{caTools::colAUC()}}.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{ci95}{Logical. Should DeLong's standard error based confidence
limits be included with the AUC estimate?}

\item{nboot}{Integer. The number of bootstrap estimates to perform.}

\item{r.seed}{Integer. The value of the random seed if
reproducibility is desired.}
}
\value{
All return a numeric scalar corresponding to the area under
the curve. For 95\% confidence intervals (\code{ci95 = TRUE}), \code{\link[=calcEmpAUC]{calcEmpAUC()}}
returns a \code{list} object with these elements:
\item{auc}{The area under the curve (empirical).}
\item{lower.limit}{lower 95\% confidence limit based on standard error AUC.}
\item{upper.limit}{upper 95\% confidence limit based on standard error AUC.}

A list containing bootstrap intervals based on the number
of bootstraps performed:
\item{auc}{The raw Pepe AUC estimate from original data.}
\item{lower.limit}{The lower CI95 of the estimate.}
\item{upper.limit}{The upper CI95 of the estimate.}
}
\description{
Calculates the area under the curve (AUC).
}
\section{Functions}{
\itemize{
\item \code{calcEmpAUC()}: Calculate the \emph{empirical} AUC, optionally with corresponding 95\%
confidence intervals according to the DeLong approach via the standard
error of the AUC estimate. This empirical AUC estimate is calculated via
the trapezoid area at each step along the x-axis of a ROC curve.

\item \code{calcPepeAUC()}: Calculate the AUC according to Margaret Pepe's book.

\item \code{calcBootAUC()}: Bootstrapped confidence intervals for the 95\% limits are calculated via
\emph{empirical} bootstrap iterations and using Pepe's AUC calculation.

}}
\examples{
n <- 20
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
calcAUC(true, pred)

# Empirical AUC
calcEmpAUC(true, pred, "disease")
calcEmpAUC(true, pred, "disease", ci95 = TRUE)  # with CI95

# Pepe's AUC
calcPepeAUC(true, pred, "disease")

# bootstrapped AUC
calcBootAUC(true, pred, "disease")
calcBootAUC(true, pred, "disease", nboot = 100, r.seed = 100)  # reproducible
}
\references{
DeLong et al. (1988) for the calculation of the Standard Error
of the Area Under the Curve (AUC) and of the difference between two AUCs.

\verb{calcPepeAUC:} M. Pepe. The Statistical
Evaluation of Medical Tests for Classification and Prediction.
}
\seealso{
\code{\link[=plotEmpROC]{plotEmpROC()}}, \code{\link[=roc_xy]{roc_xy()}}

\code{\link[=replicate]{replicate()}}, \code{\link[=plotEmpROC]{plotEmpROC()}}
}
\author{
Stu Field
}
\concept{auc}
