% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBootAUC.R
\name{calcBootAUC}
\alias{calcBootAUC}
\title{Calculate bootstrapped AUC}
\usage{
calcBootAUC(
  truth,
  predicted,
  pos.class,
  nboot = 1000,
  r.seed = sample(1000, 1)
)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{nboot}{Integer. The number of bootstrap estimates to perform.}

\item{r.seed}{Integer. The value of the random seed if
reproducibility is desired.}
}
\value{
A list containing bootstrap intervals based on the number
of bootstraps performed:
\item{auc}{The raw Pepe AUC estimate from original data.}
\item{lower.limit}{The lower CI95 of the estimate.}
\item{upper.limit}{The upper CI95 of the estimate.}
}
\description{
Confidence intervals for the 95\% limits are calculated via
\emph{empirical} bootstrap iterations and using Pepe's AUC calculation.
}
\note{
This function is used primarily in \code{\link[=plotEmpROC]{plotEmpROC()}}.
}
\examples{
withr::with_seed(10, {
  true <- sample(c("cont", "disease"), 20, replace = TRUE)
  pred <- runif(20)
})
calcBootAUC(true, pred, "disease")
calcBootAUC(true, pred, "disease", nboot = 100, r.seed = 100)  # reproducible
}
\seealso{
\code{\link[=replicate]{replicate()}}, \code{\link[=plotEmpROC]{plotEmpROC()}}

Other auc: 
\code{\link{calcAUC}()}
}
\author{
Stu Field
}
\concept{auc}
