% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcROCfit.R
\name{calcROCfit}
\alias{calcROCfit}
\title{Calculate ROC Curve Parameters}
\usage{
calcROCfit(xy, optim = c("ML", "LS"), start = c(alpha = 0.5, beta = 0.5))
}
\arguments{
\item{xy}{A data frame containing "x" (1 - tnr) and "y" (tpr)
coordinates of an empirical ROC curve. This is typically
the return value of a call to \code{\link[=roc_xy]{roc_xy()}}.}

\item{optim}{Character. One of "ML" or "LS" indicating whether
Maximum Likelihood (default) or Non-linear Least Squares should
be used in the optimization.}

\item{start}{Numeric. A \emph{named} length 2 vector of initial start
values for \eqn{\alpha} and \eqn{\beta}.
\eqn{\alpha} is the cost of a false positive and \eqn{\beta} is
the cost of a false negative.}
}
\value{
Model estimates of \eqn{\alpha} and \eqn{\beta}.
}
\description{
Use non-linear least squares to calculate the parameters
\eqn{\alpha, \beta} necessary to draw a ROC curve from
empirical data. The objective function assumed is of the form:
\deqn{y \sim 1 - (1 - x^\beta)^{(1/\alpha)}}{y ~ 1 - (1 - x^beta)^(1/alpha)}
\deqn{ \alpha \in (0, 1]}{alpha in (0, 1]}
\deqn{ \beta \in (0, 1]}{beta in (0, 1]}
\deqn{ x \in (0, 1]}{x in (0, 1]}
\cr
where the extreme x-values (0 and 1) are on the \code{y = x} equilibrium line.
}
\examples{
n    <- 15
true <- rep(c("control", "case"), each = n)
pred <- withr::with_seed(1, c(rnorm(n, 0.45, 0.2), rnorm(n, 0.65, 0.2)))

rocxy <- data.frame(roc_xy(true, pred, pos.class = "case"))
calcROCfit(rocxy)        # Max Lik
calcROCfit(rocxy, "LS")  # Least Squares

# ML with new starting values
calcROCfit(rocxy, start = c(alpha = 0.3, beta = 0.6))

# See the fit through the fitted values
ggplot2::ggplot(rocxy, ggplot2::aes(x = x, y = y)) +
  geom_roc(shape = 19, size = 2) +
  geom_rocfit(data = rocxy, col = "blue")
}
\seealso{
\code{\link[=nls]{nls()}}, \code{\link[=optim]{optim()}}, \code{\link[=plotEmpROC]{plotEmpROC()}}, \code{\link[=roc_xy]{roc_xy()}}

Other ROC: 
\code{\link{calcROCperpendicular}()},
\code{\link{createROCdata}()},
\code{\link{geom_roc}()},
\code{\link{plotEmpROC}()},
\code{\link{plotROCbootCI95}()},
\code{\link{roc_xy}()}
}
\author{
Stu Field
}
\concept{ROC}
