% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcROCperpendicular.R
\name{calcROCperpendicular}
\alias{calcROCperpendicular}
\alias{calcROCcorner}
\title{Calculate Perpendicular Distance}
\usage{
calcROCperpendicular(xy)

calcROCcorner(xy)
}
\arguments{
\item{xy}{\code{numeric(2)}. An \verb{(x, y)} coordinate on the ROC curve.}
}
\value{
\item{\code{\link[=calcROCperpendicular]{calcROCperpendicular()}}}{the perpendicular distance between the
curve and the unit line.}
\item{\code{\link[=calcROCcorner]{calcROCcorner()}}}{the distance from the ROC curve to the
top-left corner of the ROC.}
}
\description{
In receiver operating criterion curves, calculate the
optimal operating point for a binary test given a series of
\verb{(x, y)} coordinates of the ROC curve, by calculating the
\emph{perpendicular} distance between the curve and the unit (\code{x = y}) line.
The Youden distance is similar but differs in that it corresponds
to the \emph{vertical} distance between the ROC curve and the unit
line, see \code{\link[=createROCdata]{createROCdata()}}.
}
\section{Functions}{
\itemize{
\item \code{calcROCcorner()}: An alternative optimal operating point is the \emph{minimal} distance
from the ROC curve to the \verb{(0, 1)} point of the operating space,
the "top-left" corner.

}}
\note{
Differs slightly from the Youden Index (J), see \code{\link[=createROCdata]{createROCdata()}}.
}
\references{
Schisterman, et al. 2005. Optimal Cut-point and
its Corresponding Youden Index to Discriminate Individuals Using
Pooled Blood Samples. Epidemiology. 16: 73-81.
}
\seealso{
\code{\link[stats:dist]{stats::dist()}}

Other ROC: 
\code{\link{calcROCfit}()},
\code{\link{createROCdata}()},
\code{\link{geom_roc}()},
\code{\link{getROCxy}()},
\code{\link{plotEmpROC}()},
\code{\link{plotROCbootCI95}()}
}
\author{
Stu Field, Shintaro Kato
}
\concept{ROC}
