% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-roc-data.R
\name{create_roc_data}
\alias{create_roc_data}
\alias{filter_roc_data}
\alias{calc_roc_perpendicular}
\alias{calc_roc_corner}
\title{Create ROC Data Table}
\usage{
create_roc_data(
  truth,
  predicted,
  pos.class,
  do.ci = FALSE,
  cutoffs = seq(1e-05, 0.99999, length.out = length(truth)),
  include.auc = FALSE
)

filter_roc_data(
  roc_data,
  metric = c("YoudenJ", "sensitivity", "specificity", "ppv", "npv", "mcc", "perpD",
    "cutoff"),
  method = c("max", "min", "value"),
  value = NULL
)

calc_roc_perpendicular(xy)

calc_roc_corner(xy)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{do.ci}{Logical. Should binomial confidence limits be
calculated and added to the return value.}

\item{cutoffs}{Various cutoffs at which to evaluate performance.}

\item{include.auc}{Logical. Should AUC also be calculated and added as a
column to the returned data frame.}

\item{roc_data}{The result of a call to \code{\link[=create_roc_data]{create_roc_data()}}, a
\code{roc_data} object.}

\item{metric}{Character. The metric (a column of the \code{roc_data} object)
on which to filter the ROC data.}

\item{method}{\code{character(1)}. The filtering method. Can be either the
maximum ("max") or minimum ("min") of your metric, or can be a
specific value ("value") on which to filter (e.g. \code{sensitivity = 0.8}).}

\item{value}{If \code{method == "value"}, the value in \eqn{[-1, 1]} on which
to filter.}

\item{xy}{\code{numeric(2)}. An \verb{(x, y)} coordinate on the ROC curve.}
}
\value{
\code{\link[=create_roc_data]{create_roc_data()}}: An object of class \code{roc_data}, a \code{tibble},
of the ROC data evaluated at each of the cutoff evaluation points.
AUC is the area under the ROC curve, with values in \eqn{[0, 1]}.
Youden's J is a method for choosing the "optimal" cut-off based on
highest total sensitivity and specificity and is calculated via:
\deqn{
    Youden_J = sensitivity + specificity - 1
  }
For details on all other classification metrics, see \code{\link[=calc_confusion]{calc_confusion()}}.

\code{\link[=filter_roc_data]{filter_roc_data()}}: A filtered \code{roc_data} data frame.

\code{\link[=calc_roc_perpendicular]{calc_roc_perpendicular()}}: the perpendicular distance between the
curve and the unit line.

\code{\link[=calc_roc_corner]{calc_roc_corner()}}: the distance from the ROC curve to the
top-left corner of the ROC.
}
\description{
Create a data frame of ROC performance at various
predefined operating points or cutoffs, one row
per cutoff evaluation.
}
\section{Functions}{
\itemize{
\item \code{filter_roc_data()}: A filtering method for class \code{roc_data}.

\item \code{calc_roc_perpendicular()}: In receiver operating criterion curves, calculate the
optimal operating point for a binary test given a series of
\eqn{(x, y)} coordinates of the ROC curve, by calculating the
\emph{perpendicular} distance between the curve and the unit \eqn{x = y} line.
The Youden distance is similar but differs in that it corresponds
to the \emph{vertical} distance between the ROC curve and the unit
line, see \code{\link[=create_roc_data]{create_roc_data()}}.

\item \code{calc_roc_corner()}: An alternative optimal operating point is the \emph{minimal} distance
from the ROC curve to the \eqn{(0, 1)} point of the operating space,
the "top-left" corner.

}}
\note{
\code{\link[=calc_roc_perpendicular]{calc_roc_perpendicular()}} differs slightly from the
Youden's Index (J), see \code{\link[=create_roc_data]{create_roc_data()}}.
}
\examples{
n <- 200
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
roc_data <- create_roc_data(true, pred, "disease")
roc_data

create_roc_data(true, pred, "disease", do.ci = TRUE)
create_roc_data(true, pred, "disease", include.auc = TRUE)

# filter method
filter_roc_data(roc_data)
filter_roc_data(roc_data, "sensitivity", "value", 0.8)

# watch out for rounding rules when trying to return a specific value!
filter_roc_data(roc_data, "sensitivity", "value", 0.85)
filter_roc_data(roc_data, "cutoff", method = "value", value = 0.5)

# distance from unit line
calc_roc_perpendicular(c(0.5, 0.5)) # on the line
calc_roc_perpendicular(c(0, 0))     # bottom-left corner
calc_roc_perpendicular(c(0, 1))     # top-left corner
calc_roc_perpendicular(c(1, 0))     # bottom-right corner
calc_roc_perpendicular(c(1, 1))     # top-right corner
}
\references{
Schisterman, et al. 2005. Optimal Cut-point and
its Corresponding Youden Index to Discriminate Individuals Using
Pooled Blood Samples. Epidemiology. 16: 73-81.
}
\seealso{
\code{\link[=calc_roc_perpendicular]{calc_roc_perpendicular()}}, \code{\link[=calc_confusion]{calc_confusion()}}

\code{\link[stats:dist]{stats::dist()}}

Other ROC: 
\code{\link{calc_roc_fit}()},
\code{\link{geom_roc}()},
\code{\link{plot_boot_roc}()},
\code{\link{plot_emp_roc}()},
\code{\link{roc_xy}()}
}
\author{
Stu Field
}
\concept{ROC}
