% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-train.R
\name{create_train}
\alias{create_train}
\alias{is.tr_data}
\alias{plot.tr_data}
\title{Create a Training Data Object}
\usage{
create_train(data, ..., group_var, classes = NULL)

is.tr_data(data)

\method{plot}{tr_data}(x, ft, main = ft, do_pdfs = FALSE, do_log = TRUE, ...)
}
\arguments{
\item{data}{A \code{data.frame} used to create a training data set.}

\item{...}{Arguments passed to \code{\link[dplyr:filter]{dplyr::filter()}} used to subset \emph{rows}.
If passing to the S3 plot method for \code{tr_data} class objects,
additional arguments are passed to \code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDFbyGroup()}}
via the \code{...}.}

\item{group_var}{\code{character(1)}. Can be quoted or unquoted.
Must be a column name of \code{data}.}

\item{classes}{Either \code{NULL}, where no factor conversion will be
performed (default), or a string \code{character(2)} indicating \emph{first}
and \emph{second} class labels respectively. See the Details section
for more information about factor levels.}

\item{x}{A \code{tr_data} object.}

\item{ft}{\code{character(1)}. The name of a column in \code{data}
containing values to generate CDFs or PDFs.}

\item{main}{\code{character(1)}. Title for the plot. See \code{\link[ggplot2:labs]{ggplot2::ggtitle()}}.}

\item{do_pdfs}{\code{logical(1)}. Should smoothed densities PDF be plotted?}

\item{do_log}{\code{logical(1)}. Should values be log10-transformed?}
}
\value{
A \code{tibble} with an additional \code{tr_data} class.
This object contains the subset training data with a
additional attributes about the groupings and the
"response" variable.

Logical. Whether \code{data} inherits from class \code{tr_data}.
}
\description{
Generate a training data set from a original parent data set
typically via some subset of the original parent. Final groups
\emph{must} be binary, and generate a 2 factor level "response"
column used in many downstream statistical testing functions.

\code{\link[=is.tr_data]{is.tr_data()}} checks whether an object
is a \code{tr_data} class object.
}
\details{
When specifying filtering variables, the factor levels
will be ordered \emph{alphabetically} in the resulting "response"
variable unless ordering is specified by the \code{classes}
argument. This is important, for example, when performing
repeated univariate statistics where \eqn{class2 - class1},
i.e. the positive class is 2!
}
\section{Functions}{
\itemize{
\item \code{plot(tr_data)}: Plots a CDF, and optionally an accompanying smoothed PDF
for a specific feature in a "tr_data" object.

}}
\examples{
# New "tr_data" object with default factor levels
classes <- c("setosa", "versicolor")
tr <- create_train(iris, Species \%in\% classes, group_var = Species)
tr

# Getting Variables
attr(tr, "response_var")

attr(tr, "class_labels")

attr(tr, "counts")

# with re-naming factors
tr2 <- create_train(iris, Species \%in\% classes,
                    group_var = Species, classes = rev(classes))
tr2
# S3 plot method
ft <- "Sepal.Length"  # random feature
plot(tr, ft)

plot(tr, ft, cols = c("blue", "red"))

plot(tr, ft, cols = c("black", "black"))   # b/w

plot(tr, ft, do_pdfs = TRUE)

plot(tr, ft, do_pdfs = TRUE, cols = c("blue", "red"))
}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}}

\code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDFbyGroup()}}
}
\author{
Stu Field
}
