% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createROCdata.R
\name{createROCdata}
\alias{createROCdata}
\alias{filterROCdata}
\title{Create ROC Data Frame}
\usage{
createROCdata(
  truth,
  predicted,
  pos.class,
  do.ci = FALSE,
  cutoffs = seq(1e-05, 0.99999, length.out = length(truth)),
  include.auc = FALSE
)

filterROCdata(
  roc_data,
  metric = c("YoudenJ", "sensitivity", "specificity", "ppv", "npv", "mcc", "perpD",
    "cutoff"),
  method = c("max", "min", "value"),
  value = NULL
)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{do.ci}{Logical. Should binomial confidence limits be
calculated and added to the return value.}

\item{cutoffs}{Various cutoffs at which to evaluate performance.}

\item{include.auc}{Logical. Should AUC also be calculated and added as a
column to the returned data frame.}

\item{roc_data}{The result of a call to \code{\link[=createROCdata]{createROCdata()}}.
A \code{roc_data} object.}

\item{metric}{Character. The metric (a column of the \code{roc_data} object)
on which to filter the ROC data.}

\item{method}{Character. The filtering method. Can be either the
maximum ("max") or minimum ("min") of your metric, or can be a
specific value ("value") on which to filter (e.g. \code{sensitivity = 0.8}).}

\item{value}{If \code{method == "value"}, the value in \verb{[-1, 1]} on which
to filter.}
}
\value{
\code{\link[=createROCdata]{createROCdata()}}: An object of class \code{roc_data}, a data frame,
of the ROC data evaluated at each of the cutoff evaluation points.
AUC is the area under the
ROC curve, with values in \eqn{[0, 1]}.
Youden's J is a method for choosing the "optimal" cut-off based on
highest total sensitivity and specificity and is calculated via
\deqn{
    Youden_J = sensitivity + specificity - 1
  }
For details on all other classification metrics, see \code{\link[=calc_confusion]{calc_confusion()}}.

\code{\link[=filterROCdata]{filterROCdata()}}: A filtered \code{roc_data} data frame.
}
\description{
Create a data frame of ROC performance at various
predefined operating points or cutoffs, one row
per cutoff evaluation.
}
\section{Functions}{
\itemize{
\item \code{filterROCdata()}: A filtering method for class \code{roc_data}.

}}
\examples{
n <- 200
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
roc_data <- createROCdata(true, pred, "disease")
roc_data
createROCdata(true, pred, "disease", do.ci = TRUE)
createROCdata(true, pred, "disease", include.auc = TRUE)

# filter method
filterROCdata(roc_data)
filterROCdata(roc_data, "sensitivity", "value", 0.8)

# watch out for rounding rules when trying to return a specific value!
filterROCdata(roc_data, "sensitivity", "value", 0.85)
filterROCdata(roc_data, "cutoff", method = "value", value = 0.5)
}
\seealso{
\code{\link[=calcROCperpendicular]{calcROCperpendicular()}}, \code{\link[=calc_confusion]{calc_confusion()}}

Other ROC: 
\code{\link{calcROCfit}()},
\code{\link{calcROCperpendicular}()},
\code{\link{geom_roc}()},
\code{\link{plotEmpROC}()},
\code{\link{plotROCbootCI95}()},
\code{\link{roc_xy}()}
}
\author{
Mike Mehan and Stu Field
}
\concept{ROC}
