% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingData.R
\name{createTrainingData}
\alias{createTrainingData}
\alias{is.tr_data}
\alias{convert2TrainingData}
\alias{plot.tr_data}
\title{Create Training Data Object}
\usage{
createTrainingData(
  data,
  ...,
  group.var = SampleGroup,
  class1 = NULL,
  class2 = NULL
)

is.tr_data(data)

convert2TrainingData(data, group.var)

\method{plot}{tr_data}(x, apt, do.log = TRUE, main = apt, do.pdfs = FALSE, ...)
}
\arguments{
\item{data}{A \code{soma_adat} object used to create a training data set.}

\item{...}{Arguments passed to \code{\link[=filter]{filter()}} used to
subset unwanted rows from the data set.

\strong{Note:} that when specifying filtering variables, the factor levels
will be ordered \emph{alphabetically} in the resulting "Response"
variable unless ordering is specified by the \code{class1} and
\code{class2} arguments. This is important when performing
\code{\link[=calc.t]{calc.t()}}, which performs t-tests as \eqn{level2 - level1}.

If passing to the S3 plot method for objects of class \code{tr_data},
additional arguments are passed to \code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDFbyGroup()}}
via the \code{...}.}

\item{group.var}{Either a quoted or unquoted string for the \code{Response} column.
Must a column contained within \code{data}.}

\item{class1}{Character. The string to be mapped to the \emph{first}
class/level if a \emph{new} class name is desired. No mapping is performed
by default.}

\item{class2}{Character. The string to be mapped the \emph{second}
class/level if a \emph{new} class name is desired. No mapping is performed
by default.}

\item{x}{A \code{tr_data} object created via either
\code{\link[=createTrainingData]{createTrainingData()}} or \code{\link[=convert2TrainingData]{convert2TrainingData()}}.}

\item{apt}{Character. The name of a column in \code{data} containing RFU values
to generate CDFs or PDFs.}

\item{do.log}{Logical. Should RFU values be log10-transformed prior to
plotting?}

\item{main}{Character. Title for the plot. See \code{\link[ggplot2:labs]{ggplot2::ggtitle()}}.}

\item{do.pdfs}{Logical. Should smoothed densities PDF be plotted?}
}
\value{
A "grouped tibble" table (class "tbl"), with a \code{tr_data}
class added. This object contains the subset training data with a
"Response" column appended that contains the \emph{binary} groupings.

Logical. Whether \code{data} inherits from class \code{tr_data}.
}
\description{
Generate a training data set from a original parent data set
typically via some subset of the original parent. Final groups
\emph{must} be binary, and generate a 2 factor level "Response"
column used in many downstream statistical testing functions
throughout the \pkg{somaverse} environment.

\code{\link[=is.tr_data]{is.tr_data()}} checks whether an object
is a \code{tr_data} class object.

Convert a standard data frame object with a "Response"
field into a \code{tr_data} object. This is a bypass
around the typical \code{\link[=createTrainingData]{createTrainingData()}} process
if the training data was filtered via other methods.
}
\section{Functions}{
\itemize{
\item \code{plot(tr_data)}: Plots a CDF, and optionally an accompanying smoothed PDF
for a specific feature (analyte) in a "tr_data" object.

}}
\examples{
# Creating new "tr_data" object
# with default factor levels
trda <- createTrainingData(sim_test_data, gender \%in\% c("F", "M"),
                           group.var = gender)
trda$Response

# Getting Variables
dplyr::group_vars(trda)
splyr::group_labels(trda)

# with re-naming factors
trda2 <- createTrainingData(sim_test_data, gender \%in\% c("F", "M"),
                            group.var = gender,
                            class1 = "Female", class2 = "Male")
trda2$Response

# with re-naming factors AND re-ordering
trda3 <- createTrainingData(sim_test_data, gender \%in\% c("F", "M"),
                            group.var = gender,
                            class1 = "Male", class2 = "Female")
trda3$Response

# Converting -> `tr_data`
convert2TrainingData(iris, Species)               # Can be un-quoted
new_td <- convert2TrainingData(iris, "Species")   # or quoted
new_td
class(new_td)

# S3 plot method
apt <- "seq.1130.49"  # random feature
plot(trda, apt)

plot(trda, apt, cols = c("blue", "red"))

plot(trda, apt, cols = c("black", "black"))   # b/w

plot(trda, apt, do.pdfs = TRUE)

plot(trda, apt, do.pdfs = TRUE, cols = c("blue", "red"))
}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}},

\code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDFbyGroup()}}, \code{\link[=createTrainingData]{createTrainingData()}}
}
\author{
Stu Field
}
