% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-gbm.R
\name{fit_gbm}
\alias{fit_gbm}
\alias{fit_gbm.default}
\alias{fit_gbm.formula}
\title{Fit a Generalized Boosted Regression Model}
\usage{
fit_gbm(x, ...)

\method{fit_gbm}{default}(x, y, ...)

\method{fit_gbm}{formula}(formula, data, ...)
}
\arguments{
\item{x}{A \code{data.frame} containing feature data (predictors). If
using the \code{formula} method, a "Response" column should be included. The
simplest way to achieve this is a call to \code{\link[=create_train]{create_train()}}.}

\item{...}{Arguments passed to \code{\link[=gbm]{gbm()}}.}

\item{y}{\code{factor(n)}. If not passing a formula, a factor with
true class names for each sample (row) in \code{x}.}

\item{formula}{A model formula of the form:
\eqn{class ~ x1 + x2 + ...+ xn`}, (no interactions).}

\item{data}{A data frame of predictors (categorical and/or numeric).}
}
\value{
A \code{gbm} class object, as returned by \code{\link[=gbm]{gbm()}}.
}
\description{
A wrapper for fitting boosted binary regression models
for binary classification problems. Assumes a binary
"Response" in \verb{[0, 1]}.
}
\section{Methods (by class)}{
\itemize{
\item \code{fit_gbm(default)}: The S3 default method for \code{fit_gbm}.

\item \code{fit_gbm(formula)}: The S3 formula method for \code{fit_gbm}.

}}
\examples{
# formula method
model <- withr::with_seed(10, fit_gbm(Species ~ ., data = tr_iris))

# data frame method
model <- withr::with_seed(10, fit_gbm(tr_iris[, -5L], y = tr_iris$Species))
}
\seealso{
\code{\link[=gbm]{gbm()}}, \code{\link[=create_train]{create_train()}}

The `fit*()` family: 
\code{\link{fit_kknn}()},
\code{\link{fit_logistic}()},
\code{\link{fit_nb}()}
}
\author{
Stu Field
}
\concept{fit}
