% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-kknn.R
\name{fit_kknn}
\alias{fit_kknn}
\title{Fit Weighted k-Nearest Neighbor Classifier}
\usage{
fit_kknn(
  formula,
  train,
  test,
  k_neighbors = 10,
  distance = 2,
  kernel = "triangular",
  ...
)
}
\arguments{
\item{formula}{A \code{formula} class object, specifying
the model to be fitted (e.g. \eqn{response ~ x_1 + x_2 + ... + x_n}).}

\item{train}{Matrix or data frame of training set cases.}

\item{test}{Matrix or data frame of test set cases.}

\item{k_neighbors}{See \code{k} in \code{\link[kknn:kknn]{kknn::kknn()}}.}

\item{distance}{\code{numeric(1)}. Parameter of Minkowski distance.
\code{Manhattan = 1} and \code{Euclidean = 2}.}

\item{kernel}{Kernel to use. Possible choices are "rectangular" (which is
standard unweighted knn), "triangular", "epanechnikov" (or beta(2,2)),
"biweight" (or beta(3,3)), "triweight" (or beta(4,4)), "cos", "inv",
"gaussian", "rank" and "optimal".}

\item{...}{Additional arguments passed to \code{\link[kknn:kknn]{kknn::kknn()}}.}
}
\value{
A k-nearest neighbors model, as returned by \code{\link[kknn:kknn]{kknn::kknn()}},
with a \code{"Response"} variable, \code{classes}, and function parameters \code{train},
\code{k}, \code{distance}, \code{kernel}, as well as the function call (\code{call})
added as entries to the list.
}
\description{
Wrapper for fitting weighted \emph{k}-nearest neighbor classifiers.
}
\examples{
# Use fake training data from iris data set
train_idx <- sample(nrow(tr_iris), 90L)  # random 90\% training
test_df   <- tr_iris[-train_idx, ]
kknnfit   <- fit_kknn(Species ~ ., train = tr_iris[train_idx, ], test = test_df)
pos  <- get_pos_class(kknnfit)
pred <- calc_predictions(kknnfit) # internal test predictions
pred

# Confusion matrix
calc_confusion(test_df$Species, pred$prob_virginica, pos_class = pos) |>
  summary()

# plot ROC
plot_emp_roc(test_df$Species, pred$prob_virginica, pos)
}
\seealso{
\code{\link[kknn:kknn]{kknn::kknn()}}

The \verb{fit*()} family:
\code{\link{fit_gbm}()},
\code{\link{fit_logistic}()},
\code{\link{fit_nb}()}
}
\author{
Stu Field
}
\concept{fit}
