% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-nb.R
\name{fit_nb}
\alias{fit_nb}
\alias{fit_nb.default}
\alias{fit_nb.formula}
\alias{fit_nb.tr_data}
\alias{print.libml_nb}
\alias{predict.libml_nb}
\alias{plot.libml_nb}
\alias{plot.naiveBayes}
\title{Robustly Fit Naive Bayes Classifier}
\usage{
fit_nb(x, ...)

\method{fit_nb}{default}(x, y, mad = FALSE, laplace = 0, keep.data = FALSE, ...)

\method{fit_nb}{formula}(formula, data, ...)

\method{fit_nb}{tr_data}(x, ...)

\method{print}{libml_nb}(x, ...)

\method{predict}{libml_nb}(
  object,
  newdata,
  type = c("class", "posterior", "raw"),
  threshold = 1e-06,
  min.prob = NULL,
  ...
)

\method{plot}{libml_nb}(
  x,
  data,
  features,
  plot.type = c("pdf", "cdf", "log.odds"),
  x.lab = "value",
  id,
  ...
)

\method{plot}{naiveBayes}(
  x,
  data,
  features,
  plot.type = c("pdf", "cdf", "log.odds"),
  x.lab = "value",
  id,
  ...
)
}
\arguments{
\item{x}{A numeric matrix, a \code{tr_data} class objects, or a data frame
of predictors. If called from an S3 generic method (e.g.
\code{\link[=plot.libml_nb]{plot.libml_nb()}}) or \code{\link[=print.libml_nb]{print.libml_nb()}}), either a \code{libml_nb} or \code{naiveBayes} object.}

\item{...}{Additional arguments passed to the default \code{\link[=fit_nb]{fit_nb()}}
default method. Currently not used in the \code{predict} or \code{print} S3 methods,
but is used in the S3 plot method, arguments passed to
\code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDFlist()}} or \code{\link[SomaPlotr:plotPDF]{SomaPlotr::plotPDFlist()}}.}

\item{y}{A vector indicating the true classes for each sample. Ideally a
factor class object with appropriate levels.}

\item{mad}{Logical. Should non-parametric approximations be applied during
the parameter estimation procedure. See \code{Details} section.}

\item{laplace}{positive double controlling Laplace smoothing. The default
(\code{0}) disables Laplace smoothing.}

\item{keep.data}{Logical. Should the training data used to fit the model be
included in the model object? When building thousands of models, this can
become a memory issue and thus the default is \code{FALSE}.}

\item{formula}{A model formula of the form: \code{class ~ x1 + x2 + ...}
(no interactions).}

\item{data}{A data frame of predictors (categorical and/or numeric), i.e.
the ADAT used to train the model.}

\item{object}{A model object of class \code{libml_nb}.}

\item{newdata}{A \code{data.frame} with new predictors, containing at least
the model covariates (but possibly more columns than the training data).
Note that the column names of \code{newdata} are matched against the
training data ones.}

\item{type}{If \code{"class"} (default), the class name with maximal
posterior probability is returned for each sample, otherwise the
conditional \emph{a-posterior} probabilities for each class are returned.
Additionally, if called from within the S3 plot method, a character
string determining the plot type, currently either CDF or PDF (default).
Argument can be shortened and is matched.}

\item{threshold}{Value below which should be replaced. See \code{min.prob}.}

\item{min.prob}{Value indicating the minimum probability a prediction
can take. See \code{threshold} argument.}

\item{features}{An optional feature specifying which subset of model features
to plot. If missing, all features are plotted.}

\item{plot.type}{Character. A string determining the plot type, currently
either a probability density function (PDF, default), CDF, or log-odds
plots. Arguments can be shortened and is matched via \code{\link[=match.arg]{match.arg()}}.}

\item{x.lab}{Character. Optional label for the x-axis.}

\item{id}{An optional identifier of a specific sample to plot on top of
either PDFs or CDFs. This may be either a numeric index of the sample row
in the \code{data}, or its \code{rowname}. Can be \code{length(n)}.}
}
\value{
\code{libml_nb}: A naive Bayes model with robustly fit parameters.

\code{predict.libml_nb}: Depending on the \code{type} argument,
the posterior probability of a robustly estimated naive Bayes model.

\code{plot.libml_nb}, \code{plot.naiveBayes}: A plot, either a list of
PDFs/CDFs, or a log-odds plot.
}
\description{
Computes the conditional \emph{a}-posterior probabilities of a
categorical class variable given independent predictor
variables using the Bayes rule.
Parameter estimates are robustly calculated using approximations
of the error function for a Gaussian density, see \code{\link[=fit_gauss]{fit_gauss()}}.
}
\details{
When \code{mad = TRUE} (median absolute deviation), non-parametric
calculation of Bayes' parameters are estimated,
namely, \code{mu = median(x)} and \code{sd = IQR(x) / 1.349}.
That is, \code{fit_gauss(..., mad = TRUE)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{fit_nb(default)}: S3 \code{default} method for fit_nb.

\item \code{fit_nb(formula)}: S3 \code{formula} method for fit_nb.

\item \code{fit_nb(tr_data)}: S3 \code{tr_data} method for fit_nb.

}}
\section{Functions}{
\itemize{
\item \code{print(libml_nb)}: S3 print method for \code{libml_nb}.

\item \code{predict(libml_nb)}: S3 predict method for \code{libml_nb}.

\item \code{plot(libml_nb)}: A S3 plot method for \code{libml_nb}.

\item \code{plot(naiveBayes)}: Plot a \code{naiveBayes} (\code{e1071}) model object.

}}
\examples{
head(tr_iris)
# standard naive Bayes
m1 <- e1071::naiveBayes(Species ~ ., data = tr_iris) # non-robust params
m2 <- fit_nb(Species ~ ., data = tr_iris)  # formula syntax
m3 <- fit_nb(tr_iris)                      # tr_data syntax
m4 <- data.frame(tr_iris[, -5L]) |>
  fit_nb(y = tr_iris$Species)              # default syntax

# not same
identical(sapply(m1$tables, as.numeric), sapply(m2$tables, as.numeric))

# same
identical(sapply(m2$tables, as.numeric), sapply(m3$tables, as.numeric))

# Predictions
table(predict(m1, iris), iris$Species) # benchmark
table(predict(m2, iris), iris$Species) # approx same for Gaussian data; no outliers

# Plotting
plot(m2, tr_iris)
plot(m2, tr_iris, id = 50)      # sample 50 is definitely "setosa"
plot(m1, tr_iris, plot.type = "cdf")  # plot type CDF
plot(m2, tr_iris, features = "Sepal.Length", id = 70)  # 1 feature
plot(m1, tr_iris, plot.type = "cdf", lty = "longdash") # pass-through of lty
}
\references{
This function was \emph{heavily} influenced by \code{\link[e1071:naiveBayes]{e1071::naiveBayes()}}
See David Meyer <email: David.Meyer@R-project.org>.
}
\seealso{
\code{\link[=fit_gauss]{fit_gauss()}}

\code{\link[=plotLogOdds]{plotLogOdds()}}, \code{\link[SomaPlotr:plotPDF]{SomaPlotr::plotPDFlist()}}, \code{\link[SomaPlotr:plotCDF]{SomaPlotr::plotCDFlist()}}

The `fit*()` family: 
\code{\link{fit_gbm}()},
\code{\link{fit_kknn}()},
\code{\link{fit_logistic}()}
}
\author{
Stu Field
}
\concept{fit}
