% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGBM.R
\name{fitGBM}
\alias{fitGBM}
\alias{fitGBM.default}
\alias{fitGBM.formula}
\title{Fit a Generalized Boosted Regression Model}
\usage{
fitGBM(x, ...)

\method{fitGBM}{default}(x, y, ...)

\method{fitGBM}{formula}(formula, data, ...)
}
\arguments{
\item{x}{A \code{data.frame} containing feature data (predictors). If
using the \code{formula} method, a "Response" column should be included. The
simplest way to achieve this is a call to \code{\link[=createTrainingData]{createTrainingData()}}.}

\item{...}{Arguments passed to \code{\link[=gbm]{gbm()}}.}

\item{y}{Factor. If not passing a formula, a factor with
true class names for each sample.}

\item{formula}{A model formula of the form: \eqn{class ~ x1 + x2 + ...+ xn`}
(no interactions).}

\item{data}{A data frame of predictors (categorical and/or numeric).}
}
\value{
An object of class \code{gbm}, as returned by \code{\link[=gbm]{gbm()}}.
}
\description{
A wrapper for fitting boosted binary regression models
for binary classification problems. Assumes a binary
"Response" column is contained in the data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitGBM(default)}: The S3 default method for \code{fitGBM}.

\item \code{fitGBM(formula)}: The S3 formula method for \code{fitGBM}.

}}
\examples{
# default method
# Use fake training data from iris data set
model1 <- withr::with_seed(10, fitGBM(fake_iris[, -5], y = fake_iris$Response))

# formula method
model2 <- withr::with_seed(10, fitGBM(Response ~ ., data = fake_iris))
}
\seealso{
\code{\link[=gbm]{gbm()}}, \code{\link[=createTrainingData]{createTrainingData()}}

The `fit*()` family: 
\code{\link{fitGLM}()},
\code{\link{fitKKNN}()}
}
\author{
Stu Field
}
\concept{fit}
