% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\alias{fitGLM.formula}
\alias{fitGLM.data.frame}
\alias{fitGLM.tr_data}
\alias{fitGLM.matrix}
\title{Fit Multivariate Logistic Regression Model}
\usage{
fitGLM(x, ..., strip)

\method{fitGLM}{formula}(formula, ..., strip = FALSE)

\method{fitGLM}{data.frame}(x, y = NULL, strip = FALSE, ...)

\method{fitGLM}{tr_data}(x, ..., strip = FALSE)

\method{fitGLM}{matrix}(x, y, strip = FALSE, ...)
}
\arguments{
\item{strip}{Logical. Should certain entries of the model object
be stripped via \code{\link[=stripLMC]{stripLMC()}} to reduce object size?
If true, some downstream functionality is compromised, e.g. \code{\link[=summary]{summary()}} and
\code{\link[=residuals]{residuals()}}, however when iterating over 1000s of models this may be
an acceptable trade-off to limit runaway memory consumption.}

\item{formula, x, ...}{Either a formula, data frame, or matrix.
If a \code{formula} (preferred) should be a model of the
form: \eqn{class ~ x_1 + x_2 + ... + x_n}.
If a data frame (preferably a \code{tr_data} object), containing features
or predictors. If a matrix object containing ONLY predictors,
in which case \code{y} \emph{must} be passed (see examples below). Unmatched
arguments eventually be passed to \code{\link[=glm]{glm()}} via the \code{...}.}

\item{y}{Can be one of \emph{two} options:
\describe{
\item{character}{A \code{character(1)} indicating the column in \code{x}
containing the true class names.}
\item{vector}{A vector (\code{nrow(x)}) containign the true class names
for each sample. That is a vector with \emph{2} levels. If not
a factor class, it will be converted to one via \code{\link[=as.factor]{as.factor()}}.}
}}
}
\value{
A \code{glm} model object as returned by \code{\link[=glm]{glm()}}, logistic regression model.
}
\description{
A wrapper around \code{\link[=glm]{glm()}} for fitting multivariate
logistic regression models for binary classification problems.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitGLM(formula)}: S3 formula method for \code{fitGLM}.

\item \code{fitGLM(data.frame)}: S3 \code{data.frame} method for \code{fitGLM}.

\item \code{fitGLM(tr_data)}: S3 \code{tr_data} method for \code{fitGLM}.

\item \code{fitGLM(matrix)}: S3 matrix method for \code{fitGLM}.

}}
\examples{
# formula S3 method
# This is the preferred syntax
class(tr_iris)

df <- tibble::as_tibble(tr_iris)  # strip tr_data class

# tr_data S3 method:
model <- fitGLM(tr_iris)

# data frame S3 method:
model <- fitGLM(df, "Species")

# formula S3 method:
model <- fitGLM(Species ~ ., data = df)

# data frame S3 method (2):
model <- fitGLM(df[, -5L], y = df$Species)  # vector of class names

# matrix S3 method:
model <- fitGLM(as.matrix(df[, -5L]), y = df$Species)  # 'glmnet' syntax
}
\seealso{
\code{\link[=glm]{glm()}}

The `fit*()` family: 
\code{\link{fitGBM}()},
\code{\link{fitKKNN}()},
\code{\link{fit_nb}()}
}
\author{
Stu Field
}
\concept{fit}
