% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitKKNN.R
\name{fitKKNN}
\alias{fitKKNN}
\title{Fit Weighted k-Nearest Neighbor Classifier}
\usage{
fitKKNN(formula, train, test, K = 10, distance = 2, kernel = "triangular", ...)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic
description of the model to be fitted (e.g. \eqn{Response ~ terms}).}

\item{train}{Matrix or data frame of training set cases.}

\item{test}{Matrix or data frame of test set cases.}

\item{K}{Numeric. Number of neighbors considered.}

\item{distance}{Numeric. Parameter of Minkowski distance. \code{1} is
equivalent to Manhattan distance and \code{2} is equivalent to Euclidean.}

\item{kernel}{Character. Kernel to use. Possible choices are:
\describe{
\item{rectangular}{standard unweighted \code{knn}}
\item{triangular}{\eqn{1 - |u|, |u| \le 1}}
\item{epanechnikov}{beta(2, 2)}
\item{biweight}{beta(3, 3)}
\item{triweight}{beta(4, 4)}
\item{cos}{\eqn{cos(0.5\pi u)}}
\item{inv}{?}
\item{gaussian}{\eqn{exp(-0.5u^2)}}
\item{rank}{?}
\item{optimal}{?}
}}

\item{...}{Additional arguments passed to \code{\link[=kknn]{kknn()}}.}
}
\value{
A k-nearest neighbors model, as returned by \code{\link[=kknn]{kknn()}},
with \code{Response}, \code{classes}, and function parameters \code{train},
\code{K}, \code{distance}, \code{kernel}, as well as the function call (\code{call})
added as entries to the list.
}
\description{
Wrapper for fitting weighted \emph{k}-nearest neighbor classifiers.
}
\examples{
# Use fake training data from iris data set
trainIdx <- sample(nrow(fake_iris), 90)  # random 90\% training
kknnfit  <- fitKKNN(Response ~ ., train = fake_iris[trainIdx, ],
                    test = fake_iris[-trainIdx, ])
pos  <- getPositiveClass(kknnfit)
true <- fake_iris$Response[-trainIdx]   # true class names
pred <- calcPredictions(kknnfit)        # model predictions
pred

# Confusion matrix
cmat <- calc_confusion(true, pred$prob_virginica, pos.class = pos)
summary(cmat)

# plot ROC
plotEmpROC(true, pred$prob_virginica, pos)
}
\seealso{
\code{\link[=kknn]{kknn()}}

The `fit*()` family: 
\code{\link{fitGBM}()},
\code{\link{fitGLM}()}
}
\author{
Mike Mehan
}
\concept{fit}
