% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittr-package.R
\docType{package}
\name{fittr-package}
\alias{fittr}
\alias{fittr-package}
\title{fittr: Model Fitting via Machine Learning Building}
\description{
The fittr package contains general functions necessary for the fitting and performance evaluation of machine learning models. Examples include machine learning techniques, calculating KS distances, random forest, k-fold cross-validation, classification performance, and ROC plotting.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
