% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-roc.R
\name{geom_roc}
\alias{geom_roc}
\alias{geom_rocfit}
\title{Plot a ROC Curve}
\usage{
geom_roc(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  shape = NULL,
  size = 2,
  lwd = 1,
  outline = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_rocfit(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{A \code{data.frame} containing "x" and "y" coordinates corresponding
to an empirical ROC curve. This is result of a call to \code{\link[=roc_xy]{roc_xy()}}, and
corresponds to the \code{1 - "tnr"} and \code{"tpr"} values respectively.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{shape}{\code{numeric(1)}. Shape of points (between 0 and 25),
similar to \code{pch} of \code{\link[graphics:points]{graphics::points()}}. See \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{size}{\code{numeric(1)}. Size of points. Similar to \code{cex}
of \code{\link[graphics:points]{graphics::points()}}. Modifying \code{size} will not affect the plot if
\code{shape} is set to \code{NULL} (the default). See [geom_point())].

[geom_point())]: R:geom_point())}

\item{lwd}{\code{numeric(1)}. Line width (see \code{\link[=par]{par()}}).}

\item{outline}{\code{logical(1)}. Should black outlines be drawn around
the main plot line?}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:layer]{ggplot2::layer()}},
often \code{lty}, \code{shape}, \code{lwd}, etc.}
}
\description{
Create a \verb{"geom"} layer to generate a receiver operator criterion (ROC)
curve in the \pkg{ggplot2} style grammar of graphics.
Its primary input is the output of \code{\link[=roc_xy]{roc_xy()}}, and is used
primarily used in support of the wrapper \code{\link[=plot_emp_roc]{plot_emp_roc()}}.
}
\section{Functions}{
\itemize{
\item \code{geom_rocfit()}: Add a fitted line (layer) to ROC.

}}
\examples{
library(ggplot2)

# Generate dummy data
true <- rep(c("control", "disease"), each = 10)
pred <- withr::with_seed(8,
  c(rnorm(10, mean = 0.4, sd = 0.2),
    rnorm(10, mean = 0.6, sd = 0.2))
)
rocxy <- roc_xy(true, pred, "disease") |> data.frame()

# Plotting options
ggplot(rocxy, aes(x = x, y = y)) + geom_roc()
ggplot(rocxy, aes(x = x, y = y)) + geom_roc(col = "blue")
ggplot(rocxy, aes(x = x, y = y)) + geom_roc(col = "blue", outline = FALSE)

# Draw a fit line with `geom_rocfit()`
# (to add a fit-layer, you *must* pass the data argument)
ggplot(rocxy, aes(x = x, y = y)) + geom_rocfit(data = rocxy)

# Layer a fit line over a ROC curve
ggplot(rocxy, aes(x = x, y = y)) +
  geom_roc(col = "blue") +
  geom_rocfit(data = rocxy, col = "red", linetype = "dashed")

# Multiple curves can be drawn on the same plot.
# First, generate a 2nd set of dummy data
true2 <- rep(c("control", "disease"), each = 20)
pred2 <- withr::with_seed(9,
  c(rnorm(20, mean = 0.4, sd = 0.2),
    rnorm(20, mean = 0.8, sd = 0.2))
)

# Cast input to a data frame (this is required for ggplot)
rocxy2 <- roc_xy(true2, pred2, "disease") |> data.frame()

# The 2nd line can be added via standard `+` ggplot2 syntax,
# but the data argument must be passed for each geom, as each curve was
# generated from a unique dataset
ggplot() +
  geom_roc(aes(x = x, y = y), data = rocxy, col = "red") +
  geom_roc(aes(x = x, y = y), data = rocxy2, col = "blue")
}
\seealso{
\code{\link[=roc_xy]{roc_xy()}}, \code{\link[=calc_roc_fit]{calc_roc_fit()}}, \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}, \code{\link[ggplot2:layer]{ggplot2::layer()}}

Other ROC: 
\code{\link{calc_roc_fit}()},
\code{\link{create_roc_data}()},
\code{\link{plot_boot_roc}()},
\code{\link{plot_emp_roc}()},
\code{\link{roc_xy}()}
}
\author{
Stu Field, Amanda Hiser
}
\concept{ROC}
