% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cutoffs.R
\name{get-cutoffs}
\alias{get-cutoffs}
\alias{get_max_cutoff}
\alias{get_spec_cutoff}
\title{Get Distance Cutoffs}
\usage{
get_max_cutoff(truth, predicted, pos.class)

get_spec_cutoff(truth, predicted, spec, pos.class)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{spec}{Numeric. The desired specificity.}
}
\value{
\code{\link[=get_max_cutoff]{get_max_cutoff()}}: a numeric cutoff representing
the operating point at the maximal perpendicular distance
from the unit line.

\code{\link[=get_spec_cutoff]{get_spec_cutoff()}}: a numeric cutoff representing
the operating point for a given specificity.
}
\description{
For given true class names and predictions, calculate the maximal
perpendicular distance to the unit line, its corresponding
specificity, then the cutoff corresponding to that specificity.

For a given specificity, calculate the corresponding
cutoff (operating point) from a set of predictions.
}
\examples{
n <- 20
withr::with_seed(122, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
get_max_cutoff(true, pred, "disease")

# via specificity
get_spec_cutoff(true, pred, 0.4, "disease")
}
\seealso{
\code{\link[=calc_roc_perpendicular]{calc_roc_perpendicular()}}, \code{\link[=roc_xy]{roc_xy()}}
}
\author{
Stu Field
}
