% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gini.R
\name{get_gini}
\alias{get_gini}
\title{Calculate Sorted Gini Importance}
\usage{
get_gini(rf.model)
}
\arguments{
\item{rf.model}{A random forest model object (class \code{randomForest}).}
}
\value{
A \code{tibble} of the sorted importance(s). \code{MeanDecreaseAccuracy} is
the mean decrease in accuracy resulting from removing the feature.
}
\description{
Generate a table of sorted Gini importance scores for a
random forest model.
}
\examples{
# Use tr_iris training data from iris data set
rf <- withr::with_seed(101, {
  randomForest::randomForest(
    Species ~ ., data = tr_iris, importance = TRUE,
    proximity = TRUE, keep.inbag = TRUE)
})
get_gini(rf)
}
\seealso{
\code{\link[=randomForest]{randomForest()}}
}
\author{
Stu Field
}
