% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-get-model-coef.R
\name{get_model_coef-libml}
\alias{get_model_coef-libml}
\alias{get_model_coef.glm}
\alias{get_model_coef.lm}
\alias{get_model_coef.lda}
\alias{get_model_coef.libml_nb}
\alias{get_model_coef.naiveBayes}
\alias{get_model_coef.kknn}
\alias{get_model_coef.randomForest}
\alias{get_model_coef.gbm}
\alias{get_model_coef.glmnet}
\alias{get_model_coef.cv.glmnet}
\alias{get_model_coef.train}
\alias{get_model_coef.svm}
\title{Get Coefficients of a Model}
\usage{
\method{get_model_coef}{glm}(model, ...)

\method{get_model_coef}{lm}(model, ...)

\method{get_model_coef}{lda}(model, ...)

\method{get_model_coef}{libml_nb}(model, ...)

\method{get_model_coef}{naiveBayes}(model, ...)

\method{get_model_coef}{kknn}(model, ...)

\method{get_model_coef}{randomForest}(model, ...)

\method{get_model_coef}{gbm}(model, ...)

\method{get_model_coef}{glmnet}(model, lambda = NULL, ...)

\method{get_model_coef}{cv.glmnet}(model, lambda = NULL, ...)

\method{get_model_coef}{train}(model, lambda = NULL, ...)

\method{get_model_coef}{svm}(model, ...)
}
\arguments{
\item{model}{A model object, currently one of:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>  [1] get_model_coef.cv.glmnet   
#>  [2] get_model_coef.default*    
#>  [3] get_model_coef.gbm         
#>  [4] get_model_coef.glm         
#>  [5] get_model_coef.glmnet      
#>  [6] get_model_coef.kknn        
#>  [7] get_model_coef.lda         
#>  [8] get_model_coef.libml_nb    
#>  [9] get_model_coef.lm          
#> [10] get_model_coef.naiveBayes  
#> [11] get_model_coef.randomForest
#> [12] get_model_coef.svm         
#> [13] get_model_coef.train       
#> see '?methods' for accessing help and source code
}\if{html}{\out{</div>}}}

\item{...}{Additional parameters for extensibility.}

\item{lambda}{The value of the penalty parameter \code{lambda} which
can either be \code{NULL} or \code{numeric}.
If \code{NULL}, the default value depends on the underlying class:
\describe{
\item{\code{glmnet}}{The first value of lambda.}
\item{\code{cv.glmnet}}{The lambda where the cross validated error is minimized.}
\item{\code{train}}{The optimal value of lambda. When given a numeric value
of \code{lambda}, the closest value of \code{lambda} within the model will be used.}
}}
}
\value{
A named numeric vector of the coefficients of the model.
If the model is non-linear (e.g. random forest), \code{NULL}.
}
\description{
Extract the coefficients for an arbitrary model.
}
\section{Functions}{
\itemize{
\item \code{get_model_coef(glm)}: S3 method for \code{glm} models.

\item \code{get_model_coef(lm)}: S3 method for \code{lm} models.

\item \code{get_model_coef(lda)}: S3 method for \code{lda} models.

\item \code{get_model_coef(libml_nb)}: S3 method for \code{libml_nb} models.

\item \code{get_model_coef(naiveBayes)}: S3 method for \code{naiveBayes} models.

\item \code{get_model_coef(kknn)}: S3 method for \code{kknn} models.

\item \code{get_model_coef(randomForest)}: S3 method for \code{randomForest} models.

\item \code{get_model_coef(gbm)}: S3 method for \code{gbm} models.

\item \code{get_model_coef(glmnet)}: S3 method for \code{glmnet} models.

\item \code{get_model_coef(cv.glmnet)}: S3 method for \code{cv.glmnet} models.

\item \code{get_model_coef(train)}: S3 method for \code{train} models.

\item \code{get_model_coef(svm)}: S3 method for SVM models. If the model \code{kernel = linear},
coefficients are returned. Otherwise, \code{NULL}.

}}
\examples{
# set up training and test data:
iris2 <- droplevels(iris[iris$Species != "setosa", ])

# Logistic Regression
stats::glm(Species ~ ., data = iris2, family = "binomial") |>
  get_model_coef()
}
\seealso{
\code{\link[=coef]{coef()}}
}
