% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoffMax.R
\name{getCutoffMax}
\alias{getCutoffMax}
\title{Get Maximal Distance Cutoff}
\usage{
getCutoffMax(truth, predicted, pos.class)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}
}
\value{
A numeric cutoff representing the operating point
at the maximal perpendicular distance from the unit line.
}
\description{
For given true class names and predictions, calculate the maximal
perpendicular distance to the unit line, its corresponding
specificity, then the cutoff corresponding to that specificity.
}
\examples{
n <- 20
withr::with_seed(122, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
getCutoffMax(true, pred, "disease")
}
\seealso{
\code{\link[=getCutoffSpec]{getCutoffSpec()}}, \code{\link[=calc_roc_perpendicular]{calc_roc_perpendicular()}}, \code{\link[=roc_xy]{roc_xy()}}
}
\author{
Stu Field
}
