% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoffSpec.R
\name{getCutoffSpec}
\alias{getCutoffSpec}
\title{Get Cutoff for a Given Specificity}
\usage{
getCutoffSpec(truth, predicted, spec, pos.class)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{spec}{Numeric. The desired specificity.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}
}
\value{
A numeric cutoff representing the operating point
for a given specificity.
}
\description{
For a given specificity, \code{\link[=getCutoffSpec]{getCutoffSpec()}} calculates the
corresponding cutoff (operating point) from a set of predictions.
}
\examples{
n <- 20
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
getCutoffSpec(true, pred, 0.4, "disease")
}
\author{
Stu Field
}
