% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGini.R
\name{getGini}
\alias{getGini}
\title{Calculate Sorted Gini Importance}
\usage{
getGini(rf.model)
}
\arguments{
\item{rf.model}{A random forest model object, class \code{randomForest}.}
}
\value{
A \code{tibble} of the sorted importance(s). \code{MeanDecreaseAccuracy} is
the mean decrease in accuracy resulting from removing the feature.
}
\description{
Generate a table of sorted Gini importance scores for a
random forest model.
}
\examples{
# Use fake_iris training data from iris data set
library(randomForest)
rf <- withr::with_seed(101, {
  randomForest(Response ~ ., data = fake_iris, importance = TRUE,
               proximity = TRUE, keep.inbag = TRUE)
})
getGini(rf)
}
\seealso{
\code{\link[=randomForest]{randomForest()}}
}
\author{
Stu Field
}
