% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-getModelCoef.R
\name{getModelCoef-libml}
\alias{getModelCoef-libml}
\alias{getModelCoef.glm}
\alias{getModelCoef.lm}
\alias{getModelCoef.lda}
\alias{getModelCoef.robustNaiveBayes}
\alias{getModelCoef.naiveBayes}
\alias{getModelCoef.kknn}
\alias{getModelCoef.randomForest}
\alias{getModelCoef.gbm}
\alias{getModelCoef.glmnet}
\alias{getModelCoef.cv.glmnet}
\alias{getModelCoef.train}
\alias{getModelCoef.svm}
\title{Get Coefficients of a Model}
\usage{
\method{getModelCoef}{glm}(model, ...)

\method{getModelCoef}{lm}(model, ...)

\method{getModelCoef}{lda}(model, ...)

\method{getModelCoef}{robustNaiveBayes}(model, ...)

\method{getModelCoef}{naiveBayes}(model, ...)

\method{getModelCoef}{kknn}(model, ...)

\method{getModelCoef}{randomForest}(model, ...)

\method{getModelCoef}{gbm}(model, ...)

\method{getModelCoef}{glmnet}(model, lambda = NULL, ...)

\method{getModelCoef}{cv.glmnet}(model, lambda = NULL, ...)

\method{getModelCoef}{train}(model, lambda = NULL, ...)

\method{getModelCoef}{svm}(model, ...)
}
\arguments{
\item{model}{A model object, currently one of:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>  [1] getModelCoef.cv.glmnet       
#>  [2] getModelCoef.default*        
#>  [3] getModelCoef.gbm             
#>  [4] getModelCoef.glm             
#>  [5] getModelCoef.glmnet          
#>  [6] getModelCoef.kknn            
#>  [7] getModelCoef.lda             
#>  [8] getModelCoef.lm              
#>  [9] getModelCoef.naiveBayes      
#> [10] getModelCoef.randomForest    
#> [11] getModelCoef.robustNaiveBayes
#> [12] getModelCoef.svm             
#> [13] getModelCoef.train           
#> see '?methods' for accessing help and source code
}\if{html}{\out{</div>}}}

\item{...}{Additional parameters for extensibility.}

\item{lambda}{The value of the penalty parameter \code{lambda} which
can either be \code{NULL} or \code{numeric}.
If \code{NULL}, the default value depends on the underlying class:
\describe{
\item{\code{glmnet}}{The first value of lambda.}
\item{\code{cv.glmnet}}{The lambda where the cross validated error is minimized.}
\item{\code{train}}{The optimal value of lambda. When given a numeric value
of \code{lambda}, the closest value of \code{lambda} within the model will be used.}
}}
}
\value{
A named numeric vector of the coefficients of the model.
If the model is non-linear (e.g. random forest), \code{NULL}.
}
\description{
Extract the coefficients for an arbitrary model.
}
\section{Functions}{
\itemize{
\item \code{getModelCoef(glm)}: S3 method for \code{glm} models.

\item \code{getModelCoef(lm)}: S3 method for \code{lm} models.

\item \code{getModelCoef(lda)}: S3 method for \code{lda} models.

\item \code{getModelCoef(robustNaiveBayes)}: S3 method for \code{robustNaiveBayes} models.

\item \code{getModelCoef(naiveBayes)}: S3 method for \code{naiveBayes} models.

\item \code{getModelCoef(kknn)}: S3 method for \code{kknn} models.

\item \code{getModelCoef(randomForest)}: S3 method for \code{randomForest} models.

\item \code{getModelCoef(gbm)}: S3 method for \code{gbm} models.

\item \code{getModelCoef(glmnet)}: S3 method for \code{glmnet} models.

\item \code{getModelCoef(cv.glmnet)}: S3 method for \code{cv.glmnet} models.

\item \code{getModelCoef(train)}: S3 method for \code{train} models.

\item \code{getModelCoef(svm)}: S3 method for SVM models. If the model \code{kernel = linear},
coefficients are returned. Otherwise, \code{NULL}.

}}
\examples{
# set up training and test data:
iris2 <- droplevels(iris[iris$Species != "setosa", ])

# Logistic Regression
stats::glm(Species ~ ., data = iris2, family = "binomial") |>
  getModelCoef()
}
\seealso{
\code{\link[=coef]{coef()}}
}
