% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStat.R
\name{getStat}
\alias{getStat}
\title{Get a classification performance metric/statistic}
\usage{
getStat(
  x,
  which = c("Sensitivity", "Recall", "Specificity", "Precision", "PPV", "NPV",
    "Accuracy", "MCC", "Error", "F_measure", "G_mean", "Wt_Acc")
)
}
\arguments{
\item{x}{A summary object from a call to \code{\link[=calc_confusion]{calc_confusion()}},
a \code{summary.confusion_matrix} object.}

\item{which}{Character. A matched string for the test statistic desired.}
}
\value{
Double. The desired classification performance statistic.
}
\description{
Indexes into the S3 summary method for \code{confusion_matrix}
class object.
}
\details{
Optional statistics are restricted to the output of the summary
method. One of:\cr
\itemize{
\item{"Sensitivity"}
\item{"Recall"}
\item{"Specificity"}
\item{"Precision"}
\item{"PPV"}
\item{"NPV"}
\item{"Accuracy"}
\item{"Error"}
\item{"F_measure"}
\item{"G_mean"}
\item{"Wt_Acc"}
}
}
\examples{
n <- 20
withr::with_seed(22, {
  true  <- sample(c("control", "disease"), n, replace = TRUE)
  pred  <- runif(n)
})

# Summary of confusion matrix
c_mat <- calc_confusion(true, pred, pos.class = "disease") |> summary()
class(c_mat)

# pull out elements
getStat(c_mat, "Spec")
getStat(c_mat, "Sens")
getStat(c_mat, "Recall")    # same
getStat(c_mat, "Accuracy")
getStat(c_mat, "Error")     # 1 - accuracy
getStat(c_mat, "PPV")
getStat(c_mat, "NPV")
getStat(c_mat, "F_measure")
getStat(c_mat, "Wt_Acc")
}
\seealso{
\code{\link[=calc_confusion]{calc_confusion()}}
}
\author{
Stu Field
}
