% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold-cv.R
\name{kfold_cv}
\alias{kfold_cv}
\title{Calculate k-Fold Cross-Validation}
\usage{
kfold_cv(
  data,
  k,
  kknn_args = list(k_neighbors = 9L),
  model_type = c("lr", "nb", "rf", "svm", "kknn", "gbm"),
  ...
)
}
\arguments{
\item{data}{A training data set, for convenience should be
created via \code{\link[=create_train]{create_train()}}, but \emph{must} contain only the
features to be used in fitting the model and \code{"Response"} column.}

\item{k}{\code{integer(1)}. The number of folds to perform
(\emph{k}-fold cross-validation). Passed to \code{\link[wranglr:create_kfold]{wranglr::create_kfold()}}.}

\item{kknn_args}{Additional arguments as a pass-through to
\code{\link[=fit_kknn]{fit_kknn()}}, since the \code{...} is already used. Ignored unless
\code{type = "kknn"}.}

\item{model_type}{Which type of model to run.}

\item{...}{Additional arguments passed to \code{\link[wranglr:create_kfold]{wranglr::create_kfold()}}.}
}
\value{
A \code{tibble} containing model predictions, true class names,
and the fold of the samples used to make the prediction. There
should be a "test" prediction for each sample in \code{data}.
}
\description{
Perform \emph{k}-fold internal cross-validation on
a \code{tr_data} class object.
}
\examples{
# naive Bayes
# Use fake training data from iris data set
k10_nb <- kfold_cv(tr_iris, k = 10L, model_type = "nb")
k10_nb

# Boosted Regression Model
k10_b <- kfold_cv(tr_iris, k = 10L, model_type = "gbm")
k10_b

# Weighted K-Nearest-Neighbor
# Pass k_neighbors = 9 for number of neighbors in hood
k10_knn <- kfold_cv(tr_iris, k = 10L, model_type = "kknn")
k10_knn

# Random Forest
k10_rf <- kfold_cv(tr_iris, k = 10L, model_type = "rf")
k10_rf
}
\seealso{
\code{\link[=fit_nb]{fit_nb()}}, \code{\link[=fit_kknn]{fit_kknn()}}

\code{\link[randomForest:randomForest]{randomForest::randomForest()}}, \code{\link[=fit_gbm]{fit_gbm()}}, \code{\link[=fit_logistic]{fit_logistic()}}
}
\author{
Stu Field
}
