% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack-wolfe.R
\name{mack_wolfe}
\alias{mack_wolfe}
\alias{mack_wolfe.formula}
\alias{mack_wolfe.numeric}
\title{Mack-Wolfe Test}
\usage{
mack_wolfe(x, ...)

\method{mack_wolfe}{formula}(formula, data, ...)

\method{mack_wolfe}{numeric}(
  x,
  group,
  peak = "jt",
  rm_outliers = FALSE,
  alpha = 0.05,
  nperm = NULL,
  ...
)
}
\arguments{
\item{x}{\code{numeric(n)}. A numeric \emph{vector} of values.
If empty of parameters (see examples), the example
from Hollander & Wolfe is computed.}

\item{...}{Additional arguments passed to downstream S3 methods.}

\item{formula}{A formula specifying the \code{lhs} and \code{rhs}.}

\item{data}{A \code{data.frame} containing variables in the \code{formula}.}

\item{group}{\code{factor(n)}. A vector of groupings matched to \code{x}.
The factor levels are used as the grouping information.}

\item{peak}{\code{character(1)}. A string corresponding to the
desired peak factor level.
If \code{peak = "jt"} (default) a JT-test is performed.
If \code{peak = NULL}, the "peak unknown" version of the
Mack-Wolfe test is performed.}

\item{rm_outliers}{\code{logical(1)}. Should statistical outliers
(\eqn{6 * mad} \emph{and} \eqn{5x}) be removed? See \code{\link[=remove_outliers]{remove_outliers()}}.}

\item{alpha}{\code{numeric(1)}. The desired significance level.}

\item{nperm}{\code{integer(1)}. The number of Monte-Carlo simulations
to perform. If \code{NULL}, a p-value approximation is
used (if \code{length(x) > 10}).}
}
\value{
Returns a \code{mack_wolfe} class object:
\item{Ap:}{The Mack-Wolfe test statistic (if peak known)}
\item{Astar:}{The Normal approximation of the test statistic}
\item{p_value:}{The \strong{two-sided} p-value associated with the
Normal approximation (\code{\link[=pnorm]{pnorm()}})}
\item{Acrit:}{The critical value for the z-distribution at the
requested significance level (\code{alpha})}
\item{alpha:}{The significance level at which the critical
value was evaluated}
\item{peak:}{The peak as determined by test statistic}
\item{groups:}{The factor levels of supplied \code{group} parameter.}
}
\description{
Calculates the Mack-Wolfe test on a single vector
of numeric data. The Jonckheere-Terpstra (JT) test
or Jonckheere trend test is a special case of the
Mack-Wolfe where the peak is set to one of the ends.
}
\examples{
x <- c(36.0, 33.6, 26.9, 35.8, 30.1, 31.2, 35.3,   # g1
       39.9, 29.1, 43.4,                           # g2
       44.6, 54.4, 48.2, 55.7, 50,                 # g3
       53.8, 53.9, 62.5, 46.6,                     # g4
       44.3, 34.1, 35.7, 35.6,                     # g5
       31.7, 22.1, 30.7)                           # g6
months <- c(g1 = "Jan-Feb",
            g2 = "Mar-Apr",
            g3 = "May-Jun",
            g4 = "Jul-Aug",
            g5 = "Sep-Oct",
            g6 = "Nov-Dec")
group  <- rep(months, c(7, 3, 5, 4, 4, 3))
group  <- factor(group, months)
peak   <- "Jul-Aug"

mack_wolfe(x, group, peak)

# run example from text (same as above)
mack_wolfe()

# use Monte-Carlo permutation to estimate p-value
mack_wolfe(NA, nperm = 100L)

# S3 formula method
df <- data.frame(var = x, gr = group)
mack_wolfe(var ~ gr, data = df, peak = peak)
}
\references{
Myles Hollander and Douglas A. Wolfe (1973).
\emph{Nonparametric Statistical Methods}. New York: John Wiley & Sons.
Pages 115-120, 215.
}
\author{
Stu Field, Michael Mehan
}
