% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mackwolfe.R
\name{mackwolfe}
\alias{mackwolfe}
\title{Mack-Wolfe Test}
\usage{
mackwolfe(
  x,
  group,
  peak = NULL,
  rm.outliers = FALSE,
  alpha = 0.05,
  nperm = NULL,
  verbose = interactive()
)
}
\arguments{
\item{x}{Numeric. A \emph{vector} of values. If missing, the example
from Hollander & Wolfe is computed.}

\item{group}{Factor. A vector of groupings matched to \code{x}.
The factor levels are used as the grouping information.}

\item{peak}{Character. A string corresponding to the desired peak factor
level (\emph{not} an integer, a string). If \code{peak = "jt"} a
JT-test is performed. If \code{peak = NULL}, the "peak unknown" version of
the Mack-Wolfe test is performed.}

\item{rm.outliers}{Logical. Should statistical outliers
(\eqn{6 * mad} \emph{and} \eqn{5x}) be removed? See \code{\link[=get_outliers]{get_outliers()}} for
outlier definition.}

\item{alpha}{Numeric. The desired significance level.}

\item{nperm}{Integer. The number of Monte-Carlo simulations to perform.
If \code{NULL}, a p-value approximation is used (ensure \code{length(x) > 10}).}

\item{verbose}{Logical. Should checks of process be printed to the console?}
}
\value{
Returns a list of:
\item{Ap:}{The Mack-Wolfe test statistic (peak known)}
\item{Astar:}{The Normal approximation of the test statistic}
\item{p.value:}{The \strong{two-sided} p-value associated with the
Normal approximation (\code{\link[=pnorm]{pnorm()}})}
\item{Acrit:}{The critical value for the z-distribution at the
requested significance level (\code{alpha})}
\item{alpha:}{The significance level at which the critical
value was evaluated}
\item{peak:}{The peak as determined by test statistic}
\item{groups:}{The factor levels of supplied \code{group} parameter.}
}
\description{
Calculates the Mack-Wolfe test on a single vector of numeric data.
The Jonckheere-Terpstra (JT) test or Jonckheere Trend test is a special
case of the Mack-Wolfe test where the peak is set to one of the end points.
}
\examples{
# pass no data -> run example
mackwolfe()

# Use Monte-Carlo Permutation P-value estimation
mackwolfe(nperm = 100)
}
\references{
Myles Hollander and Douglas A. Wolfe (1973).
\emph{Nonparametric Statistical Methods}. New York: John Wiley & Sons.
Pages 115-120, 215.
}
\seealso{
\code{\link[=calc.mackwolfe]{calc.mackwolfe()}}
}
\author{
Stu Field, Michael Mehan
}
