% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Common Parameters in \pkg{fittr}}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{cutoff}{Numeric. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{main}{Character. Optional string for the plot title.}

\item{y.lab}{Character. Optional label for the y-axis.}

\item{alpha}{Numeric in \verb{[0, 1]}. The color transparency level.
See also \code{\link[ggplot2:reexports]{ggplot2::alpha()}}.}

\item{color}{Character or integer. Specify the colors for
lines, points, bar, box, or ROC.}

\item{col}{Character or integer. Specify the colors for
lines, points, bar, box, or ROC.}

\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.
Should contain a "Response" column/field indicating the dependent
response variable, often the grouping variable. The \verb{response =}
variable can also be specified to override the default.}

\item{apts}{Character vector of \code{AptNames} to use. If \code{NULL} (default),
all analyte RFU fields in \code{data} are used.}

\item{response}{Character. The column name to use as the
response grouping. Expected to be a factor with 2 levels.
If not a factor, \code{response} will be coerced to a
factor (i.e. alphabetical order). Level 2 of the factor is
considered the \code{disease/case} group, thus positive values
for the statistic indicate up-regulation in \code{Level2} compared \code{Level1}.}

\item{bh}{Logical. Indicating whether returned p-values will be
adjusted for false discovery by "Benjamini & Hochberg"
correction (\code{TRUE}) or by Storey's q-value correction.}

\item{n}{Integer. The number of rows to show in the S3 print method.}

\item{file}{A file name to write table output, typically a \verb{*.csv}.}

\item{nboot}{Integer. The number of bootstrap estimates to perform.}

\item{r.seed}{Integer. The value of the random seed if
reproducibility is desired.}
}
\value{
A list object of class \verb{c(stat_table, *_table)} containing:
\item{stat.table}{A data frame (table) of the statistical results (see below)}
\item{test}{The name of the statistical test performed, e.g. "KS-test".}
\item{call}{The direct call made the \verb{calc.*()}.}
\item{paired}{Logical. Whether the values are paired (if applicable).}
\item{response}{The "Response" column used in the statistical test.}
\item{counts}{Group counts split by the "Response" column (if applicable).}
\item{log}{Logical. Whether the RFU values were log10-transformed.}
\item{data.frame}{The name of the data frame containing the RFU data.}
\item{data.dim}{The dimensions of the data frame used for testing.}

The \code{stat.table} contains:
\itemize{
\item \strong{Test Statistic}: KS-distance, t-statistic, U, W, H, r, rho, tau, etc.
\item \strong{Signed Statistic}: Signed test statistic (if applicable).
\item \strong{p.value}: Unadjusted p-values for the test statistic.
\item \strong{fdr or q.value}: FDR adjustment to p-values (either BH or Storey).
\item \strong{p.bonferroni}: Bonferroni corrected p-value.
\item \strong{rank}: Ranking by p-value.
\item \strong{\code{loCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
\item \strong{\code{hiCI95}}: Lower 95\% confidence interval (\code{\link[=calc.cor]{calc.cor()}} Pearson only).
}
}
\description{
The parameters below are commonly used throughout
the \pkg{fittr} package.
}
