% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bayes-boundary.R
\name{plot_bayes_boundary}
\alias{plot_bayes_boundary}
\title{Plot a Naive Bayes Decision Boundary}
\usage{
plot_bayes_boundary(data, pos_class, res = 50L)
}
\arguments{
\item{data}{A \code{data.frame} object containing 3 columns:
\enumerate{
\item \strong{F1}: values for the first feature (x-axis).
\item \strong{F2}: values for the second feature (y-axis).
\item \strong{class}: vector of the response classes as a factor.
}}

\item{pos_class}{\code{character(1)}. Name of the "positive" or "event" class.}

\item{res}{\code{integer(1)}. The resolution for the plot. Higher
resolutions require more computation time.}
}
\description{
Plots the bivariate curved decision boundary for a naive
Bayes classifier/model for \emph{two} (bivariate) features.
}
\examples{
data <- data.frame(F1    = tr_iris$Petal.Length,
                   F2    = tr_iris$Sepal.Length,
                   class = tr_iris$Species)
head(data)
plot_bayes_boundary(data, pos_class = "virginica")
}
\seealso{
\code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}}, \code{\link[=fit_nb]{fit_nb()}}
}
\author{
Stu Field
}
