% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boot-roc.R
\name{plot_boot_roc}
\alias{plot_boot_roc}
\title{Plot a ROC with CI95}
\usage{
plot_boot_roc(
  truth,
  predicted,
  pos_class,
  shade_color = "black",
  nboot = 1000,
  r_seed = 101,
  add = FALSE,
  ...
)
}
\arguments{
\item{truth}{\code{character(n)} or \code{factor(n)}. A vector of true class names.
In most instances you will have to also pass a \code{pos_class} argument
defining the positive/event class.}

\item{predicted}{\code{numeric(n)}. A numeric vector of class probabilities.}

\item{pos_class}{\code{character(1)}. Name of the "positive" or "event" class.}

\item{shade_color}{The color for the bootstrap shaded region.
Passed as a \code{fill} argument to downstream \code{ggplot2} machinery.}

\item{nboot}{\code{integer(1)}. The number of bootstrap estimates to perform.}

\item{r_seed}{`integer(1). The value of the random seed if
reproducibility is desired.}

\item{add}{Logical. Should a plotting layer be added to an existing plot?}

\item{...}{Additional arguments passed to \code{\link[=geom_roc]{geom_roc()}}, e.g. \verb{color =}.}
}
\description{
Plots a ROC curve with bootstrapped 95\% confidence interval
boundary overlay.
}
\examples{
n <- 75
true <- rep(c("control", "disease"), each = n)
pred <- withr::with_seed(1, c(rnorm(n, 0.2, 0.3), rnorm(n, 0.8, 0.3)))
plot_boot_roc(true, pred, pos_class = "disease", nboot = 200, color = "blue")

# add layer
pred2 <- withr::with_seed(1, c(rnorm(n, 0.2, 0.3), rnorm(n, 0.5, 0.3)))
plot_boot_roc(true, pred, pos_class = "disease", nboot = 200,
              shade_color = "blue", color = "blue") +
plot_boot_roc(true, pred2, pos_class = "disease", nboot = 200,
              shade_color = "green", color = "red", add = TRUE)
}
\seealso{
Other ROC: 
\code{\link{calc_roc_fit}()},
\code{\link{create_roc_data}()},
\code{\link{geom_roc}()},
\code{\link{plot_emp_roc}()},
\code{\link{roc_xy}()}
}
\author{
Stu Field, Amanda Hiser
}
\concept{ROC}
