% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-log-odds.R
\name{plot_log_odds}
\alias{plot_log_odds}
\title{Create a Log-Odds Plot}
\usage{
plot_log_odds(
  truth,
  predicted,
  pos_class,
  cutoff = 0.5,
  y_lab = NULL,
  max_prob = NULL,
  scramble = FALSE
)
}
\arguments{
\item{truth}{\code{character(n)} or \code{factor(n)}. A vector of true class names.
In most instances you will have to also pass a \code{pos_class} argument
defining the positive/event class.}

\item{predicted}{\code{numeric(n)}. A numeric vector of class probabilities.}

\item{pos_class}{\code{character(1)}. Name of the "positive" or "event" class.}

\item{cutoff}{\code{numeric(1)}. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{y_lab}{\code{character(1)}. Optional label for the y-axis.}

\item{max_prob}{\code{numeric(1)}. Experimental. Maximum probability
value cutoff for the log-odds plot. Removes extreme samples
from the plot to avoid distorting x-axis.}

\item{scramble}{\code{logical(1)}. Should values be randomized to avoid
monotonically decreasing probability scores (aesthetic)?}
}
\description{
Plot the log-odds, \deqn{log(Prob / (1 - Prob))} for each sample.
See \code{Section} for note about extreme probabilities.
}
\section{Extreme probabilities}{

extreme values in \verb{[0, 1]} are thresholded at
\code{.Machine$double.eps^0.5}, or \verb{[1.490116119e-08, 0.9999999851]}
to restrict the x-axis and avoid \code{Inf} values in
log-odds space (\verb{0/1}).
}

\examples{
n <- 20
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
plot_log_odds(true, pred, "disease")
plot_log_odds(true, pred, "disease", scramble = TRUE)
}
\author{
Stu Field
}
