% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Create Manhattan Plot}
\usage{
plot_manhattan(
  data,
  x.lab = "Feature",
  type = c("t.test", "log2fc", "ks.test"),
  as_pvalue = FALSE
)
}
\arguments{
\item{data}{A training data \code{tr_data} object.}

\item{x.lab}{\code{character(1)}. Set the x-axis label.}

\item{type}{\code{character(1)}. The type measure used to
evaluate expression change.
One of: "log2fc", "t.test", or "ks.test". Pattern is matched.}

\item{as_pvalue}{\code{logical(1)}. Should p-values be
plotted in linear space or log10-space?}
}
\description{
Creates a Manhattan plot of the differences or log-ratios
of the training set for each analyte.
}
\examples{
tr <- create_train(sim_adat,
                   class_response \%in\% c("control", "disease"),
                   group.var = class_response,
                   classes = c("control", "disease"))

# Various options for plotting
plot_manhattan(tr, type = "t")
plot_manhattan(tr, type = "log2")
plot_manhattan(tr, type = "ks")
plot_manhattan(tr, type = "t", as_pvalue = TRUE)
}
\author{
Stu Field
}
