% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-uni-contrasts.R
\name{plot_uni_contrasts}
\alias{plot_uni_contrasts}
\title{Contrast 2 Univariate Tables}
\usage{
plot_uni_contrasts(
  x,
  y,
  cutoff = 0.05/nrow(x),
  ident = FALSE,
  label_size = 2.5
)
}
\arguments{
\item{x}{The first univariate tibble to contrast (x-axis).}

\item{y}{The second univariate tibble to contrast (y-axis).}

\item{cutoff}{\code{numeric(1)}. A \code{p-value} cutoff for comparison of tables.}

\item{ident}{\code{logical(1)}. Should the points beyond the cutoff be identified?}

\item{label_size}{\code{numeric(1)}. The size for the labels if \code{ident = TRUE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Compare 2 \code{univariate} tables from any two analyses
from the \code{\link[=calc_univariate]{calc_univariate()}} function.
}
\examples{
a <- calc_univariate(mtcars, var = "vs")
b <- calc_univariate(mtcars, var = "mpg", "lm")
plot_uni_contrasts(a, b, ident = TRUE, cutoff = 0.005)
}
\author{
Stu Field
}
