% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBivariateBayesBoundary.R
\name{plotBivariateBayesBoundary}
\alias{plotBivariateBayesBoundary}
\title{Plot a Naive Bayes Decision Boundary}
\usage{
plotBivariateBayesBoundary(data, pos.class, res = 50, main = NULL)
}
\arguments{
\item{data}{A \code{data.frame} object containing 3 columns:
\enumerate{
\item \strong{F1}: values for the first feature (x-axis).
\item \strong{F2}: values for the second feature (y-axis).
\item \strong{class}: vector of the response classes as a factor.
}}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{res}{Integer. The resolution for the plot. Higher
resolutions require more computation time.}

\item{main}{Character. Optional string for the plot title.}
}
\description{
Plots the bivariate curved decision boundary for a naive
Bayes classifier/model for \emph{two} features.
}
\examples{
data <- data.frame(F1    = tr_iris$Petal.Length,
                   F2    = tr_iris$Sepal.Length,
                   class = tr_iris$Species)
head(data)
plotBivariateBayesBoundary(data, pos.class = "virginica")
}
\seealso{
\code{\link[=geom_contour]{geom_contour()}}, \code{\link[=fit_nb]{fit_nb()}}
}
\author{
Stu Field
}
