% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContrastTables.R
\name{plotContrastTables}
\alias{plotContrastTables}
\title{Contrast 2 Analysis Tables}
\usage{
plotContrastTables(
  x,
  y,
  ident = FALSE,
  label.size = 2.5,
  main = "Table p-value Contrasts",
  cutoff = 0.05/nrow(x)
)
}
\arguments{
\item{x}{The first \code{stat.table} object to contrast (x-axis).}

\item{y}{The second \code{stat.table} object to contrast (y-axis).}

\item{ident}{Logical. Should the points beyond the cutoff be identified?}

\item{label.size}{The size for the labels if \code{ident = TRUE}.}

\item{main}{Optional. A title for the plot.}

\item{cutoff}{Numeric. Bonferroni corrects \code{p-value} cutoff to base
the comparison of the two tables.}
}
\value{
A comparison plot of 2 analyses.
}
\description{
Compare 2 \code{stat.table} objects from any two analyses
from the \verb{calc.*} family of functions, determining the
common significant analytes.
Each table should have a "p-value" column and the row names of each
table \emph{must} be proper \verb{"AptNames"} (e.g. \code{GeneId.SeqId} or \code{seq.SeqId}).
}
\examples{
a <- calc.t(log10(sim_test_data), response = "class_response")
b <- calc.ks(log10(sim_test_data), response = "class_response")
plotContrastTables(a, b, ident = TRUE, cutoff = 0.005, main = "Check this out!")
}
\author{
Stu Field
}
