% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmpROC.R
\name{plotEmpROC}
\alias{plotEmpROC}
\title{Plot Empirical ROC Curve}
\usage{
plotEmpROC(
  truth,
  predicted,
  pos.class,
  auc = TRUE,
  add = 0L,
  boot.auc = FALSE,
  adj = c(0, 0),
  auc.pos = c(0.5, 0.5),
  auc.shift = 1.25,
  auc.label = NULL,
  auc.size = 5,
  shape = NULL,
  size = 2,
  cutoff = 0.5,
  cutoff.size = 5,
  cutoff.shape = 23,
  col = 1,
  ci95 = TRUE,
  lwd = 2,
  outline = TRUE,
  boxes = TRUE,
  box.alpha = 0.35,
  debug = FALSE,
  plot.fit = FALSE,
  do.grid = TRUE,
  pch = deprecated(),
  pch.cex = deprecated(),
  pch.cutoff = deprecated(),
  auc.cex = deprecated(),
  cutoff.cex = deprecated(),
  ... = deprecated()
)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{auc}{Logical. Should the AUC be printed on the plot?}

\item{add}{Integer. The position in the plotting stack indicating where to
add the ROC curve relative to an existing plot.
Zero indexing is used, thus \code{add = 0L} (default) refers to a new plot
and \code{add = 1L} will add a ROC layer to the existing plot and
correctly position the \code{AUC} annotations shifted accordingly.}

\item{boot.auc}{Logical. Should bootstrap confidence intervals of AUC
be calculated?}

\item{adj}{Numeric in \verb{[0, 1]}. Coordinates that are used to align
the AUC text.}

\item{auc.pos}{Numeric in \verb{[0, 1]} indicating where to place the AUC
text. Must be of \code{length == 2L}, indicating the x-axis and y-axis positions,
respectively. By default, the AUC will be placed slightly to the right
of center of the plot.}

\item{auc.shift}{The vertical (downward) shift between AUC text values if
multiple ROCs are plotted.}

\item{auc.label}{Character string. Adds an additional label to the AUC text
(Default: "AUC = 0.99", with label "Extra Text": "Extra Text AUC = 0.99").
Must be added individually to each plot.}

\item{auc.size}{Character size of the AUC text.}

\item{shape}{Numeric. Shape of points. Must be a number between 0 and 25,
similar to \code{pch} of \code{\link[graphics:points]{graphics::points()}}. See \code{\link[=geom_point]{geom_point()}}.}

\item{size}{Numeric. Size of points. Similar to \code{cex} of \code{\link[graphics:points]{graphics::points()}}.
Modifying \code{size} will not affect the plot if \code{shape} is set to
\code{NULL} (the default). See [geom_point())].

[geom_point())]: R:geom_point())}

\item{cutoff}{The decision cutoff, aka operating point, for the
\emph{positive} classes. By default, the operating point is set to 0.5.
Alternatively, choosing a negative number calculates the cutoff
corresponding to the maximum \emph{perpendicular} distance between the
curve and the unit diagonal. The CI95 of the operating point (boxes)
can be omitted by setting \code{cutoff = NA}.}

\item{cutoff.size}{Numeric in \verb{[0, 1]}, the character size for
the cutoff point symbol.}

\item{cutoff.shape}{Numeric. The point symbol used for the cutoff point
plotted on the ROC. Defaults to a diamond (\code{23}).}

\item{col}{Character or integer. Specify the colors for
lines, points, bar, box, or ROC.}

\item{ci95}{Logical. Should any confidence intervals be plotted?}

\item{lwd}{Line width. See \code{\link[=par]{par()}}.}

\item{outline}{Logical. Should black outlines be drawn around the main plot line?}

\item{boxes}{Logical. Should confidence interval boxes be drawn (showing the
joint CI95 of the sensitivity and specificity confidence intervals) at the
cutoff point?}

\item{box.alpha}{Numeric in \verb{[0, 1]}, the shading transparency for
the confidence interval box at a given cutoff.}

\item{debug}{Logical. Should debugging mode be activated? When activated,
annotates the values of the plotting steps at each cutoff, prints the
"positive class", prints the prediction data to the console, and
various other internal objects useful for debugging.}

\item{plot.fit}{Logical/Character. Should a maximum-likelihood
(or least-squares if ML fails) fit of the curve be plotted?
If \code{plot.fit = "both"}, a fit will be \emph{added} to the ROC.
If \code{plot.fit = TRUE}, only the fit will be plotted, without the ROC.}

\item{do.grid}{Logical. Should grid lines be added to the ROC curve?}

\item{pch, pch.cex, pch.cutoff, cutoff.cex, auc.cex}{Deprecated. Now use
\code{shape}, \code{size}, \code{cutoff.shape}, \code{cutoff.size}, or \code{auc.size}.}

\item{...}{Deprecated. Maintained for backwards compatibility.}
}
\value{
A ROC curve is plotted and its corresponding AUC is returned.
}
\description{
Plotting function to generate a receiver operator criterion
(ROC) curve for binary data and binary classifiers. This function is
a wrapper around \code{\link[=geom_roc]{geom_roc()}}, with window dressing for commonly used
aesthetics and annotations.
}
\examples{
true <- rep(c("control", "disease"), each = 50)
pred <- withr::with_seed(8,
          c(rnorm(50, mean = 0.4, sd = 0.2),    # control predictions
            rnorm(50, mean = 0.6, sd = 0.2))    # disease predictions
        )

plotEmpROC(true, pred, pos.class = "disease", col = "dodgerblue")
plotEmpROC(true, pred, pos.class = "disease", ci95 = TRUE, boxes = FALSE, col = "red")

plotEmpROC(true, pred, pos.class = "disease", ci95 = FALSE, shape = 21, col = "green")
plotEmpROC(true, pred, pos.class = "disease", boot.auc = TRUE, col = "royalblue")

plotEmpROC(true, pred, pos.class = "disease", plot.fit = "both", col = "purple")
plotEmpROC(true, pred, pos.class = "disease", plot.fit = TRUE, ci95 = FALSE,
           auc = FALSE, cutoff = NA, col = 2, lwd = 4) # curve only, no cutoff

# Debugging with `debug = TRUE` displays
# curve points to the console
plotEmpROC(true, pred, pos.class = "disease", debug = TRUE, col = "firebrick3")

# Multiple curves can be drawn on the same plot
true2 <- rep(c("control", "disease"), each = 50)
pred2 <- withr::with_seed(8,
          c(rnorm(50, mean = 0.5, sd = 0.3),
            rnorm(50, mean = 0.7, sd = 0.4))
        )
plotEmpROC(true, pred, pos.class = "disease", col = "firebrick3") +
  plotEmpROC(true2, pred2, pos.class = "disease", col = "forestgreen", add = 1)
}
\seealso{
\code{\link[=create_roc_data]{create_roc_data()}}, \code{\link[=roc_xy]{roc_xy()}}, \code{\link[=geom_roc]{geom_roc()}}

Other ROC: 
\code{\link{calc_roc_fit}()},
\code{\link{create_roc_data}()},
\code{\link{geom_roc}()},
\code{\link{plotROCbootCI95}()},
\code{\link{roc_xy}()}
}
\author{
Michael R. Mehan, Stu Field
}
\concept{ROC}
