% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLogOdds.R
\name{plotLogOdds}
\alias{plotLogOdds}
\title{Create a Log-Odds Plot}
\usage{
plotLogOdds(
  truth,
  predicted,
  pos.class,
  cutoff = 0.5,
  main = NULL,
  y.lab = NULL,
  max.prob = NULL,
  scramble = FALSE
)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{cutoff}{Numeric. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{main}{Character. Optional string for the plot title.}

\item{y.lab}{Character. Optional label for the y-axis.}

\item{max.prob}{Numeric. Maximum probability value cutoff for
the log-odds plot. Removes extreme samples from the plot
to avoid distorting x-axis.}

\item{scramble}{Logical. Should values be randomized to avoid
monotonically decreasing probability scores (aesthetic)?}
}
\description{
Plot the log-odds, \deqn{log(Prob / (1 - Prob))} for each sample.
See \code{Details} for note about extreme probabilities.
}
\details{
\strong{Note:} extreme values in \verb{[0, 1]} are thresholded at
\code{.Machine$double.eps^0.5}, or \verb{[1.490116119e-08, 0.9999999851]}
to restrict the x-axis and avoid \code{Inf} values in log-odds space (\verb{0/1}).
}
\examples{
n <- 20
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
plotLogOdds(true, pred, "disease")
plotLogOdds(true, pred, "disease", scramble = TRUE)
}
\author{
Stu Field
}
