% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROCbootCI95.R
\name{plotROCbootCI95}
\alias{plotROCbootCI95}
\title{Plot a ROC with CI95}
\usage{
plotROCbootCI95(
  truth,
  predicted,
  pos.class,
  shade.color = "black",
  nboot = 1000,
  r.seed = 101,
  add = FALSE,
  ...
)
}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{shade.color}{The color for the bootstrap shaded region.
Passed as a \code{fill} argument to downstream \code{ggplot2} machinery.}

\item{nboot}{Integer. The number of bootstrap estimates to perform.}

\item{r.seed}{Integer. The value of the random seed if
reproducibility is desired.}

\item{add}{Logical. Should a plotting layer be added to an existing plot?}

\item{...}{Additional arguments passed to \code{\link[=geom_roc]{geom_roc()}}, e.g. \verb{color =}.}
}
\description{
Plots a ROC curve with bootstrapped 95\% confidence interval
boundary overlay.
}
\examples{
n <- 75
true <- rep(c("control", "disease"), each = n)
pred <- withr::with_seed(1, c(rnorm(n, 0.2, 0.3), rnorm(n, 0.8, 0.3)))
plotROCbootCI95(true, pred, pos.class = "disease", nboot = 200, color = "blue")

# add layer
pred2 <- withr::with_seed(1, c(rnorm(n, 0.2, 0.3), rnorm(n, 0.5, 0.3)))
plotROCbootCI95(true, pred, pos.class = "disease", nboot = 200,
                shade.color = "blue", color = "blue") +
plotROCbootCI95(true, pred2, pos.class = "disease", nboot = 200,
                shade.color = "green", color = "red", add = TRUE)
}
\seealso{
Other ROC: 
\code{\link{calcROCfit}()},
\code{\link{calcROCperpendicular}()},
\code{\link{createROCdata}()},
\code{\link{geom_roc}()},
\code{\link{plotEmpROC}()},
\code{\link{roc_xy}()}
}
\author{
Stu Field, Amanda Hiser
}
\concept{ROC}
