% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-stat.R
\name{pull_stat}
\alias{pull_stat}
\title{Pull a Classification Performance Metric/Statistic}
\usage{
pull_stat(
  x,
  which = c("Sensitivity", "Recall", "Specificity", "Precision", "PPV", "NPV",
    "Accuracy", "MCC", "AUC", "Brier", "Error", "F_measure", "G_mean", "Wt_Acc")
)
}
\arguments{
\item{x}{A \code{summary_confusion_matrix} class object, from a
call to \code{summary(calc_confusion(...))}.}

\item{which}{\code{character(1)}. Matched string for the test desired statistic.}
}
\value{
\code{double(1)}. The classification performance statistic.
}
\description{
Indexes into the S3 summary method for \code{confusion_matrix}
class object.
}
\details{
Optional statistics are restricted to the output of the summary
method. One of:\cr
\itemize{
\item{"Sensitivity"}
\item{"Recall"}
\item{"Specificity"}
\item{"Precision"}
\item{"PPV"}
\item{"NPV"}
\item{"Accuracy"}
\item{"Error"}
\item{"F_measure"}
\item{"G_mean"}
\item{"Wt_Acc"}
}
}
\examples{
n <- 20
withr::with_seed(22, {
  true  <- sample(c("control", "disease"), n, replace = TRUE)
  pred  <- runif(n)
})

# Summary of confusion matrix
c_mat <- calc_confusion(true, pred, pos.class = "disease") |> summary()
class(c_mat)

# pull out elements
pull_stat(c_mat, "Spec")
pull_stat(c_mat, "Sens")
pull_stat(c_mat, "Recall")    # same
pull_stat(c_mat, "Accuracy")
pull_stat(c_mat, "Error")     # 1 - accuracy
pull_stat(c_mat, "PPV")
pull_stat(c_mat, "NPV")
pull_stat(c_mat, "F_measure")
pull_stat(c_mat, "Wt_Acc")
}
\seealso{
\code{\link[=calc_confusion]{calc_confusion()}}
}
\author{
Stu Field
}
