% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-calcPredictions.R, R/s3-getModelCoef.R,
%   R/s3-getTableStats.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{calcPredictions}
\alias{getModelCoef}
\alias{getTableStats}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{globalr}{\code{\link[globalr:s3-generics]{calcPredictions}}, \code{\link[globalr:s3-generics]{getModelCoef}}, \code{\link[globalr:s3-generics]{getTableStats}}}
}}

