% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc-xy.R
\name{roc_xy}
\alias{roc_xy}
\title{ROC Curve Coordinates}
\usage{
roc_xy(truth, predicted, pos.class)
}
\arguments{
\item{truth}{\code{character(n)} or \code{factor(n)}. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{\code{numeric(n)}. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}
}
\value{
A matrix containing the \code{x} and \code{y} coordinates for the
ROC curve. A matrix is preferred over a data frame for speed of indexing
while iterating over the rows and having to convert between classes.
Downstream code will often convert to data frame while the main AUC
functionality prefers a matrix.
}
\description{
Calculate the the \verb{(x, y)} coordinates of an empirical ROC curve.
}
\details{
This algorithm was adapted from the one in Fawcett (2006)
to account to a more accurate step calculation of indices
with ties. The original paper suggests moving along the
diagonal when tied according to the \emph{expected} sensitivity
and specificity, however this does not account for ties
that occur \emph{within} the same class, in which case a walk
along the edge of the "unknown" box is the correct decision. In
this algorithm, a step in the diagonal only occurs if there is
a tie \emph{and} the current class name differs from the previous.
Otherwise, a full step occurs in the appropriate direction, up
for positive classes, right for negative classes.
}
\examples{
n <- 25
withr::with_seed(22, {
  true <- sample(c("control", "disease"), n, replace = TRUE)
  pred <- runif(n)
})
xy <- roc_xy(true, pred, "disease")
xy

# simple plotting
ggplot2::ggplot(data.frame(xy), ggplot2::aes(x = x, y = y)) +
  geom_roc(outline = FALSE, shape = 19)
}
\references{
Fawcett, Tom. 2006. An introduction to ROC analysis. Pattern
Recognition Letters. 27:861-874.
}
\seealso{
\code{\link[=plot_emp_roc]{plot_emp_roc()}}, \code{\link[=create_roc_data]{create_roc_data()}}

Other ROC: 
\code{\link{calc_roc_fit}()},
\code{\link{create_roc_data}()},
\code{\link{geom_roc}()},
\code{\link{plot_boot_roc}()},
\code{\link{plot_emp_roc}()}
}
\author{
Stu Field
}
\concept{ROC}
