% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-calc-predictions.R
\name{s3-calc_predictions}
\alias{s3-calc_predictions}
\alias{calc_predictions.libml_nb}
\alias{calc_predictions.naiveBayes}
\alias{calc_predictions.randomForest}
\alias{calc_predictions.gbm}
\alias{calc_predictions.svm}
\alias{calc_predictions.glm}
\alias{calc_predictions.kknn}
\title{Calculate Model Predictions}
\usage{
\method{calc_predictions}{libml_nb}(model, newdata, cutoff = 0.5, ...)

\method{calc_predictions}{naiveBayes}(model, newdata, cutoff = 0.5, ...)

\method{calc_predictions}{randomForest}(model, newdata = NULL, cutoff = 0.5, ...)

\method{calc_predictions}{gbm}(model, newdata, cutoff = 0.5, ...)

\method{calc_predictions}{svm}(model, newdata, cutoff = 0.5, ...)

\method{calc_predictions}{glm}(model, newdata, cutoff = 0.5, ...)

\method{calc_predictions}{kknn}(model, newdata = NULL, cutoff = 0.5, ...)
}
\arguments{
\item{model}{A model object. Currently one of:
\itemize{
\item \code{glm} (for logistic regression)
\item \code{libml_nb}
\item \code{naiveBayes}
\item \code{randomForest}
\item \code{svm}
\item \code{gbm}
\item \code{kknn}
}}

\item{newdata}{The test set (\code{data.frame}) containing
(features) corresponding to the model parameters.
For some models, if \code{newdata = NULL}, training
(or out-of-bag) predictions are returned.}

\item{cutoff}{\code{numeric(1)}. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{...}{Used for extensibility of downstream S3 methods.}
}
\value{
A \code{data.frame} with predicted class and class probabilities for
each row in \code{newdata}:
\item{\code{pred_class}}{predicted class for each new observation.
If there are more than two classes, the class with
the \emph{highest} predicted probability. Otherwise, class
predictions are returned based on the value of \code{cutoff}.}
\item{\verb{prob_*}}{probability of belonging to class \code{*} for each new observation.}
\item{\code{pred_linear}}{for linear models, the linear predictor for
each new observation.}
}
\description{
Calculate the test set predictions of a given model and test set.
\itemize{
\item For the \code{randomForest} model, if \code{newdata = NULL} (default),
the out-of-bag sample predictions are returned, thus no test data
are required.
\item For KKNN models, the test set predictions are maintained
with the model itself, thus you should \emph{not} pass \code{test}.
\item For \code{randomForest} and \code{gbm} models, there are out-of-bag samples
that are used for predictions if \code{newdata = NULL}.
}
}
\section{Functions}{
\itemize{
\item \code{calc_predictions(libml_nb)}: S3 method for "robust" Naive Bayes models.

\item \code{calc_predictions(naiveBayes)}: S3 method for Naive Bayes models.

\item \code{calc_predictions(randomForest)}: S3 method for Random Forest models.

\item \code{calc_predictions(gbm)}: S3 method for Gradient Boosted Tree models.

\item \code{calc_predictions(svm)}: S3 method for Support Vector Machines.

\item \code{calc_predictions(glm)}: S3 method for Logistic Regression via \code{glm}.

\item \code{calc_predictions(kknn)}: S3 method for Weighted k-Nearest Neighbor.

}}
\examples{
# Use training data from iris data set:
train <- head(tr_iris, -3L)
test  <- tibble::as_tibble(tail(tr_iris, 3L))

# Logistic Regression
lr <- fit_logistic(Species ~ ., data = train)
calc_predictions(lr, test)
calc_predictions(lr, test, cutoff = 0.33)

# Naive Bayes
nb <- fit_nb(Species ~ ., data = train)
calc_predictions(nb, test)
calc_predictions(nb, test, cutoff = 0.01)

# Random Forest
rf <- withr::with_seed(101, {
  randomForest::randomForest(Species ~ ., data = train,
                             importance = TRUE, proximity = TRUE,
                             keep.inbag = TRUE)
})
head(calc_predictions(rf))        # out-of-bag
calc_predictions(rf, test)
calc_predictions(rf, test, cutoff = 0.39)

# Generalized Boosted Regression Model
gb <- fit_gbm(Species ~ ., data = train)
calc_predictions(gb, test)
calc_predictions(gb, test, 0.48)

# Support Vector Machines
sm <- e1071::svm(Species ~ ., data = train, probability = TRUE)
calc_predictions(sm, test)
calc_predictions(sm, test, 0.35)

# KKNN
# test data passed during fitting:
kknn <- fit_kknn(Species ~ ., train = train, test = test, k_neighbors = 10)
calc_predictions(kknn)                 # do NOT pass test data
calc_predictions(kknn, cutoff = 0.48)
}
\seealso{
\code{\link[=fit_nb]{fit_nb()}}, \code{\link[=randomForest]{randomForest()}}, \code{\link[=fit_logistic]{fit_logistic()}}, \code{\link[=fit_kknn]{fit_kknn()}}
}
\author{
Stu Field
}
