% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-calcPredictions.R
\name{s3-calcPredictions}
\alias{s3-calcPredictions}
\alias{calcPredictions.robustNaiveBayes}
\alias{calcPredictions.naiveBayes}
\alias{calcPredictions.randomForest}
\alias{calcPredictions.gbm}
\alias{calcPredictions.svm}
\alias{calcPredictions.glm}
\alias{calcPredictions.kknn}
\alias{calcPredictions.soma_lognet}
\title{Calculate Model Predictions}
\usage{
\method{calcPredictions}{robustNaiveBayes}(model, newdata, cutoff = 0.5, ...)

\method{calcPredictions}{naiveBayes}(model, newdata, cutoff = 0.5, ...)

\method{calcPredictions}{randomForest}(model, newdata = NULL, cutoff = 0.5, ...)

\method{calcPredictions}{gbm}(model, newdata, cutoff = 0.5, ...)

\method{calcPredictions}{svm}(model, newdata, cutoff = 0.5, ...)

\method{calcPredictions}{glm}(model, newdata, cutoff = 0.5, ...)

\method{calcPredictions}{kknn}(model, newdata = NULL, cutoff = 0.5, ...)

\method{calcPredictions}{soma_lognet}(model, newdata, cutoff = 0.5, ...)
}
\arguments{
\item{model}{A model object. Currently one of:
\itemize{
\item \code{glm} (for logistic regression)
\item \code{robustNaiveBayes}
\item \code{naiveBayes}
\item \code{randomForest}
\item \code{svm}
\item \code{gbm}
\item \code{kknn}
\item \code{soma_lognet}
}}

\item{newdata}{The test set (\code{data.frame}) containing protein
(features) corresponding to the model parameters.
For some models, if \code{newdata = NULL}, training (or out-of-bag)
predictions are returned.}

\item{cutoff}{Numeric. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{...}{Used for extensibility of downstream S3 methods.}
}
\value{
A \code{data.frame} with predicted class and class probabilities for
each row in \code{newdata}:
\item{\code{pred_class}}{predicted class for each new observation.
If there are more than two classes, the class with
the \emph{highest} predicted probability. Otherwise, class
predictions are returned based on the value of \code{cutoff}.}
\item{\verb{prob_*}}{probability of belonging to class \code{*} for each new observation.}
\item{\code{pred_linear}}{for linear models, the linear predictor for
each new observation.}
}
\description{
Calculate the test set predictions of a given model and test set.
\itemize{
\item For the \code{randomForest} model, if \code{newdata = NULL} (default),
the out-of-bag sample predictions are returned, thus no test data
are required.
\item For KKNN models, the test set predictions are maintained
with the model itself, thus you should \emph{not} pass \code{test}.
\item For \code{randomForest} and \code{gbm} models, there are out-of-bag samples
that are used for predictions if \code{newdata = NULL}.
}
}
\section{Functions}{
\itemize{
\item \code{calcPredictions(robustNaiveBayes)}: S3 method for "robust" Naive Bayes models.

\item \code{calcPredictions(naiveBayes)}: S3 method for Naive Bayes models.

\item \code{calcPredictions(randomForest)}: S3 method for Random Forest models.

\item \code{calcPredictions(gbm)}: S3 method for Gradient Boosted Tree models.

\item \code{calcPredictions(svm)}: S3 method for Support Vector Machines.

\item \code{calcPredictions(glm)}: S3 method for Logistic Regression via \code{glm}.

\item \code{calcPredictions(kknn)}: S3 method for Weighted k-Nearest Neighbor.

\item \code{calcPredictions(soma_lognet)}: S3 method for Logistic Regression Elastic Net models.

}}
\examples{
# Use training data from iris data set:
train <- head(tr_iris, -3L)
test  <- tibble::as_tibble(tail(tr_iris, 3L))

# Logistic Regression
lr <- fitGLM(Species ~ ., data = train)
calcPredictions(lr, test)
calcPredictions(lr, test, cutoff = 0.33)

# Naive Bayes
nb <- robustNaiveBayes(Species ~ ., data = train)
calcPredictions(nb, test)
calcPredictions(nb, test, cutoff = 0.01)

# Random Forest
rf <- withr::with_seed(101, {
  randomForest::randomForest(Species ~ ., data = train,
                             importance = TRUE, proximity = TRUE,
                             keep.inbag = TRUE)
})
head(calcPredictions(rf))        # out-of-bag
calcPredictions(rf, test)
calcPredictions(rf, test, cutoff = 0.39)

# Generalized Boosted Regression Model
gb <- fitGBM(Species ~ ., data = train)
calcPredictions(gb, test)
calcPredictions(gb, test, 0.48)

# Support Vector Machines
sm <- e1071::svm(Species ~ ., data = train, probability = TRUE)
calcPredictions(sm, test)
calcPredictions(sm, test, 0.35)

# KKNN
# test data passed during fitting:
kknn <- fitKKNN(Species ~ ., train = train, test = test, K = 10)
calcPredictions(kknn)                 # do NOT pass test data
calcPredictions(kknn, cutoff = 0.48)
}
\seealso{
\code{\link[=robustNaiveBayes]{robustNaiveBayes()}}, \code{\link[=randomForest]{randomForest()}}, \code{\link[=glm]{glm()}}, \code{\link[=kknn]{kknn()}}
}
\author{
Stu Field
}
