% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-features.R
\name{select_features}
\alias{select_features}
\title{Select Model Features}
\usage{
select_features(model, .data)
}
\arguments{
\item{model}{A model with an S3 \code{\link[=getModelFeatures]{getModelFeatures()}} method. See description
in \code{\link[=getModelFeatures]{getModelFeatures()}}.}

\item{.data}{A \code{data.frame}, typically containing a test data
of samples to to subset down to the minimal set of features.}
}
\value{
A \code{data.frame}-like object (depending on class of \code{.data}) subset
to only the variables/features contained in the \code{model}.
}
\description{
Subsets a data frame to only the model predictor
variables (columns), aka the "model frame", for a given a \emph{model}.
Similar to \code{\link[=model.frame]{model.frame()}}, except a \emph{model} is passed rather than
a \emph{formula}.
}
\examples{
# set up training and test data:
idx   <- sample.int(nrow(tr_iris), size = 90L)
train <- tr_iris[idx, ]
test  <- tibble::as_tibble(tr_iris[-idx, ])

lr <- fit_logistic(train)
select_features(lr, train)

select_features(lr, test)

\dontrun{
  select_features(lr, test[, -2L]) # throws error; missing feature
}

# Generalized Boosted Regression Model
gb <- fit_gbm(Species ~ ., data = train)
select_features(gb, test)

# Support Vector Machines
sm <- e1071::svm(Species ~ ., data = train, probability = TRUE)
select_features(sm, test)

# KKNN
# note: test data passed during fitting
kknn <- fit_kknn(Species ~ ., train = train, test = test, K = 10)
select_features(kknn, test)
}
\author{
Stu Field
}
