% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Select Model Features}
\usage{
selectFeatures(model, .data)
}
\arguments{
\item{model}{A model with an S3 \code{\link[=getModelFeatures]{getModelFeatures()}} method. See description
in \code{\link[=getModelFeatures]{getModelFeatures()}}.}

\item{.data}{A \code{data.frame}, typically containing a test data
of samples to to subset down to the minimal set of features.}
}
\value{
A \code{data.frame}-like object (depending on class of \code{.data}) subset
to only the variables/features contained in the \code{model}.
}
\description{
Subsets a data frame to only the model predictor
variables (columns), aka the "model frame", for a given a \emph{model}.
Similar to \code{\link[=model.frame]{model.frame()}}, except a \emph{model} is passed rather than
a \emph{formula}.
}
\examples{
# set up training and test data:
idx   <- sample.int(nrow(fake_iris), size = 90L)
train <- fake_iris[idx, ]
test  <- fake_iris[-idx, ]

lr <- fitGLM(Response ~ ., data = train)
selectFeatures(lr, train)

selectFeatures(lr, test)

\dontrun{
  selectFeatures(lr, test[, -2L]) # throws error; missing feature
}

# Generalized Boosted Regression Model
gb <- fitGBM(Response ~ ., data = train)
selectFeatures(gb, test)

# Support Vector Machines
sm <- e1071::svm(Response ~ ., data = train, probability = TRUE)
selectFeatures(sm, test)

# KKNN
# note: test data passed during fitting
test <- tail(fake_iris, 3L)
kknn <- fitKKNN(Response ~ ., train = train, test = test, K = 10)
selectFeatures(kknn, test)
}
\author{
Stu Field
}
