% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripLMC.R
\name{stripLMC}
\alias{stripLMC}
\title{Strip Linear Model Components}
\usage{
stripLMC(x)
}
\arguments{
\item{x}{A model object of class \code{"lm"} or \code{"glm"}.}
}
\value{
A stripped down (size) \code{"lm"} or \code{"glm"} object.
For \code{"lm"} class, removes or sets -> 0 these elements:
- \code{model}
- \code{fitted}
- \code{assign}
- \code{effects}
- \code{xlevels}
For \code{"glm"} class, removes or sets to \code{0} these elements:
- \code{linear.predictors}
- \code{terms}
- \code{formula}
- \code{family} (\code{family}, \code{link}, \code{linkfun}, and \code{linkinv} remain)
}
\description{
Strips linear model (\code{"lm"} or \code{"glm"}) object components
down the bare essentials to reduce size. Used typically
when building thousands of models, e.g. during
cross-validation to manage memory.
}
\examples{
fat_lm <- function() {
  junk <- runif(1e5)
  stats::lm(vs ~ ., data = mtcars)
}
fat_glm <- function() {
  junk <- runif(1e5)   # excessive junk in envir
  stats::glm(vs ~ wt + disp, data = mtcars, family = "binomial")
}

# LM
lobstr::obj_size(fat_lm())
lobstr::obj_size(stripLMC(fat_lm()))

# GLM
lobstr::obj_size(fat_glm())
lobstr::obj_size(stripLMC(fat_glm()))
}
\seealso{
\code{\link[=glm]{glm()}}, \code{\link[=lm]{lm()}}, \code{\link[=summary]{summary()}}, \code{\link[=residuals]{residuals()}}
}
\author{
Stu Field
}
