% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr_iris}
\alias{tr_iris}
\alias{fake_iris}
\title{Iris Data Set as Training Data}
\format{
An object of class \code{tr_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 5 columns.
}
\usage{
tr_iris
}
\description{
Generate a fake \code{tr_data} object based on the
\link[datasets:iris]{datasets::iris} data set for use in
examples sections of the libml package. First, the
\code{versicolor} class observations are removed to make
the result binary, then half (\code{n = 50}) of the observations
are class scrambled, and the feature data is "jittered"
to add some noise. This was because the original
iris data set has a distinct decision boundary across the
4 features, making classification examples "too" perfect.
The format is a 4 feature + 1 class variable "tibble"
object recast as a \code{tr_data} object.
Contains 50 setosa and 50 virginica samples.
}
\note{
this is the same as \code{fake_iris} pre-existing object.
}
\examples{
\dontrun{
  ?iris
}

# print
tr_iris
}
\references{
Fisher, R. A. (1936) The use of multiple
measurements in taxonomic problems. \emph{Annals of Eugenics}, \strong{7},
Part II, 179-188.
The data were collected by Anderson, Edgar (1935). The irises of
the Gaspe Peninsula, \emph{Bulletin of the American Iris Society}, \strong{59}, 2-5.
}
\keyword{datasets}
