% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-univariate.R
\name{calc_univariate}
\alias{calc_univariate}
\title{Create Table of Univariate Results}
\usage{
calc_univariate(data, var, test = c("t.test", "lm"), save.test = FALSE)
}
\arguments{
\item{data}{A \code{soma_adat} object containing data for analysis.}

\item{var}{Character. A response variable, the column name in \code{data}.}

\item{test}{Character. A statistical test to run.
See above for currently supports tests.}

\item{save.test}{Logical. Save a list column of the tests/models. Recommended
default value of \code{FALSE} as storing the models lot of memory.
The \code{model} term in an \code{lm} object, for example, can be very
memory-intensive.}
}
\value{
A \code{tibble}. Common columns across tests include analyte meta data:
\itemize{
\item \code{AptName}
\item \code{SeqId}
\item \code{Target}
\item \code{EntrezGeneSymbol}
\item \code{UniProt}
}

Common p-value columns are:
\itemize{
\item \code{p.value}
\item \code{FDR}
\item \code{p.bonferroni}
\item \code{rank}
}

Test-specific statistics include the following:
\itemize{
\item \code{t.test} returns the t statistic, \code{t}.
\item \code{lm} returns the intercept, slope, and t statistic of the slope,
\code{intercept}, \code{slope}, and \code{t.slope} respectively.
}
}
\description{
Iterates over analytes to compute univariate tests for each one.
Currently supports:
\itemize{
\item{t-tests for binary endpoints}
\item{linear models for continuous endpoints}
}
}
\examples{
reg_feat <- attr(sim_test_data, "sig_feats")$reg
class_feat <- attr(sim_test_data, "sig_feats")$class
small_dat <- sim_test_data[, c("gender", "age", reg_feat, class_feat)] |>
  log10() |>
  center_scale()
tts <- calc_univariate(small_dat, "gender", "t.test")
lms <- calc_univariate(small_dat, "age", "lm")
}
\author{
Stu Field
}
