% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-univariate.R
\name{calc_univariate}
\alias{calc_univariate}
\title{Create Table of Univariate Results}
\usage{
calc_univariate(data, var, test = c("t.test", "lm", "ks", "log2fc"), ...)
}
\arguments{
\item{data}{A \code{data.frame} object containing data for analysis.}

\item{var}{\code{character(1)}. A response variable, a column in \code{data}.}

\item{test}{\code{character(1)}. A statistical test to run.
See above for currently supports tests.}

\item{...}{Additional parameters passed to the statistic
function defined in \code{test}.}
}
\value{
A \code{tibble} of features and univariate test results.
Common columns are:
\itemize{
\item \code{p_value}
\item \code{FDR}
\item \code{p_bonferroni}
\item \code{rank}
}

Test-specific statistics include the following:
\itemize{
\item \code{t.test} returns the t statistic, \code{t}.
\item \code{lm} returns the intercept, slope, and t statistic of the slope,
\code{intercept}, \code{slope}, and \code{t_slope} respectively.
\item \code{log2fc} returns the log2-fold-change of the ratio of group medians.
}
}
\description{
Iterates over features to compute univariate tests for each
given an appropriate response variable.
Currently supports:
\itemize{
\item{t-tests for binary endpoints}
\item{linear models for continuous endpoints}
\item{log2FC: ratio of group medians)}
}
}
\examples{
calc_univariate(mtcars, "vs")

calc_univariate(mtcars, "vs", "ks")

calc_univariate(mtcars, "mpg", "lm")

calc_univariate(mtcars, "vs", "log2")
}
\author{
Stu Field
}
