% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-kknn.R
\name{fit_kknn}
\alias{fit_kknn}
\title{Fit Weighted k-Nearest Neighbor Classifier}
\usage{
fit_kknn(
  formula,
  train,
  test,
  k = 10,
  distance = 2,
  kernel = "triangular",
  ...
)
}
\arguments{
\item{formula}{A \code{formula} class object, specifying
the model to be fitted (e.g. \eqn{response ~ x_1 + x_2 + ... + x_n}).}

\item{train}{Matrix or data frame of training set cases.}

\item{test}{Matrix or data frame of test set cases.}

\item{k}{Number of neighbors considered.}

\item{distance}{\code{numeric(1)}. Parameter of Minkowski distance.
\code{Manhattan = 1} and \code{Euclidean = 2}.}

\item{kernel}{Character. Kernel to use. Possible choices are:
\describe{
\item{rectangular}{standard unweighted \code{knn}}
\item{triangular}{\eqn{1 - |u|, |u| \le 1}}
\item{epanechnikov}{beta(2, 2)}
\item{biweight}{beta(3, 3)}
\item{triweight}{beta(4, 4)}
\item{cos}{\eqn{cos(0.5\pi u)}}
\item{inv}{?}
\item{gaussian}{\eqn{exp(-0.5u^2)}}
\item{rank}{?}
\item{optimal}{?}
}}

\item{...}{Additional arguments passed to \code{\link[=kknn]{kknn()}}.}
}
\value{
A k-nearest neighbors model, as returned by \code{\link[=kknn]{kknn()}},
with a \code{"Response"} variable, \code{classes}, and function parameters \code{train},
\code{k}, \code{distance}, \code{kernel}, as well as the function call (\code{call})
added as entries to the list.
}
\description{
Wrapper for fitting weighted \emph{k}-nearest neighbor classifiers.
}
\examples{
# Use fake training data from iris data set
trainIdx <- sample(nrow(tr_iris), 90)  # random 90\% training
kknnfit  <- fit_kknn(Species ~ ., train = tr_iris[trainIdx, ],
                    test = tr_iris[-trainIdx, ])
pos  <- get_pos_class(kknnfit)
true <- tr_iris$Species[-trainIdx]   # true class names
pred <- calc_predictions(kknnfit)    # model predictions
pred

# Confusion matrix
calc_confusion(true, pred$prob_virginica, pos_class = pos) |>
  summary()

# plot ROC
plot_emp_roc(true, pred$prob_virginica, pos)
}
\seealso{
\code{\link[=kknn]{kknn()}}

The `fit*()` family: 
\code{\link{fit_gbm}()},
\code{\link{fit_logistic}()},
\code{\link{fit_nb}()}
}
\author{
Stu Field
}
\concept{fit}
