% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGBM.R
\name{fitGBM}
\alias{fitGBM}
\alias{fitGBM.default}
\alias{fitGBM.formula}
\title{Fit a Generalized Boosted Regression Model}
\usage{
fitGBM(x, ...)

\method{fitGBM}{default}(x, y, ...)

\method{fitGBM}{formula}(formula, data, ...)
}
\arguments{
\item{x}{A \code{data.frame} containing feature data (predictors). If
using the \code{formula} method, a "Response" column should be included. The
simplest way to achieve this is a call to \code{\link[=create_train]{create_train()}}.}

\item{...}{Arguments passed to \code{\link[=gbm]{gbm()}}.}

\item{y}{Factor. If not passing a formula, a factor with
true class names for each sample.}

\item{formula}{A model formula of the form:
\eqn{class ~ x1 + x2 + ...+ xn`}, (no interactions).}

\item{data}{A data frame of predictors (categorical and/or numeric).}
}
\value{
An object of class \code{gbm}, as returned by \code{\link[=gbm]{gbm()}}.
}
\description{
A wrapper for fitting boosted binary regression models
for binary classification problems. Assumes a binary
"Response" in \verb{[0, 1]}.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitGBM(default)}: The S3 default method for \code{fitGBM}.

\item \code{fitGBM(formula)}: The S3 formula method for \code{fitGBM}.

}}
\examples{
# formula method
model <- withr::with_seed(10, fitGBM(Species ~ ., data = tr_iris))

# data frame method
model <- withr::with_seed(10, fitGBM(tr_iris[, -5L], y = tr_iris$Species))
}
\seealso{
\code{\link[=gbm]{gbm()}}, \code{\link[=create_train]{create_train()}}

The `fit*()` family: 
\code{\link{fitGLM}()},
\code{\link{fitKKNN}()}
}
\author{
Stu Field
}
\concept{fit}
