% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\alias{fitGLM.formula}
\alias{fitGLM.data.frame}
\alias{fitGLM.matrix}
\title{Fit Multivariate Logistic Regression Model}
\usage{
fitGLM(x, ..., strip)

\method{fitGLM}{formula}(formula, ..., strip = FALSE)

\method{fitGLM}{data.frame}(x, y = NULL, strip = FALSE, ...)

\method{fitGLM}{matrix}(x, y, strip = FALSE, ...)
}
\arguments{
\item{strip}{Logical. Should certain entries of the model object
be stripped via \code{\link[=stripLMC]{stripLMC()}} to reduce object size?
If true, some downstream functionality is compromised, e.g. \code{\link[=summary]{summary()}} and
\code{\link[=residuals]{residuals()}}, however when iterating over 1000s of models this may be
an acceptable trade-off to limit runaway memory consumption.}

\item{formula, x, ...}{Either a formula, data frame, or matrix.
If a \code{formula} (preferred) should be a model of the
form: \eqn{class ~ x_1 + x_2 + ... + x_n}.
If a data frame (preferably a "tr_data" object), containing features
or predictors. If a column named "Response" is present, it is assumed to be
the column of \code{true} classes and is re-coded to be \code{y} (and \code{y = NULL}
is allowed). Alternatively, a matrix object containing ONLY predictors,
in which case \code{y} \emph{must} be passed (see examples below). Unmatched
arguments eventually make their way to \code{\link[=glm]{glm()}}.}

\item{y}{Can be one of \emph{three} options:
\describe{
\item{NULL}{Assumes a column named "Response" is in \code{x}
which contains true class names.}
\item{factor}{A vector indicating the true class names for each
sample. That is, a factor class object with \emph{2} levels.}
\item{character}{A string indicating the column in \code{x} containing
the true class names.}
}}
}
\value{
A \code{glm} model object as returned by \code{\link[=glm]{glm()}}, logistic regression model.
}
\description{
A wrapper around \code{\link[=glm]{glm()}} for fitting multivariate
logistic regression models for binary classification problems.
}
\section{Methods (by class)}{
\itemize{
\item \code{fitGLM(formula)}: S3 formula method for \code{fitGLM}.

\item \code{fitGLM(data.frame)}: S3 data frame method for \code{fitGLM}.

\item \code{fitGLM(matrix)}: S3 matrix method for \code{fitGLM}.

}}
\examples{
# formula S3 method
# This is the preferred syntax
model <- fitGLM(Response ~ ., data = fake_iris)

# data frame S3 method: 1
model <- fitGLM(fake_iris)   # assume "Response" is present

# data frame S3 method: 2
model <- fitGLM(fake_iris[, -5], y = fake_iris$Response)  # pass vector of class names

# data frame S3 method: 3
model <- fitGLM(fake_iris, y = "Response")  # pass string containing class names

# matrix S3 method (*must* pass `y` here)
model <- fitGLM(as.matrix(fake_iris[, -5]), y = fake_iris$Response)
}
\seealso{
\code{\link[=glm]{glm()}}

The `fit*()` family: 
\code{\link{fitGBM}()},
\code{\link{fitKKNN}()}
}
\author{
Stu Field
}
\concept{fit}
