% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Common Parameters in \pkg{libml}}
\arguments{
\item{truth}{Character or factor. A vector of true class names.
In most instances you will have to also pass a \code{pos.class} argument
defining the positive/event class.}

\item{predicted}{Numeric. A numeric vector of class probabilities.}

\item{pos.class}{Character. Name of the "positive" or "event" class.}

\item{cutoff}{Numeric. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{main}{Character. Optional string for the plot title.}

\item{y.lab}{Character. Optional label for the y-axis.}

\item{alpha}{Numeric in \verb{[0, 1]}. The color transparency level.
See also \code{\link[ggplot2:reexports]{ggplot2::alpha()}}.}

\item{color}{Character or integer. Specify the colors for
lines, points, bar, box, or ROC.}

\item{col}{Character or integer. Specify the colors for
lines, points, bar, box, or ROC.}

\item{data}{A \code{tibble} or \code{data.frame} object containing data for analysis.
Should often contain a "Response" column indicating the
response variable, often the grouping variable.}

\item{feats, apts}{Character vector of features, usually column names
of a data frame.}

\item{response}{Character. The column name to use as the
response grouping. Expected to be a factor with 2 levels.
If not a factor, \code{response} will be coerced to a
factor (i.e. alphabetical order). Level 2 of the factor is
considered the \code{disease/case} group, thus positive values
for the statistic indicate up-regulation in \code{Level2} compared \code{Level1}.}

\item{nboot}{Integer. The number of bootstrap estimates to perform.}

\item{r.seed}{Integer. The value of the random seed if
reproducibility is desired.}
}
\description{
The parameters below are commonly used throughout
the \pkg{libml} package.
}
