% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Common Parameters in \pkg{libml}}
\arguments{
\item{truth}{\code{character(n)} or \code{factor(n)}. A vector of true class names.
In most instances you will have to also pass a \code{pos_class} argument
defining the positive/event class.}

\item{predicted}{\code{numeric(n)}. A numeric vector of class probabilities.}

\item{pos_class}{\code{character(1)}. Name of the "positive" or "event" class.}

\item{cutoff}{\code{numeric(1)}. A cutoff for the decision/operating point,
predictions above which are considered the \emph{positive} class.}

\item{main}{\code{character(1)}. Optional string for the plot title.}

\item{y_lab}{\code{character(1)}. Optional label for the y-axis.}

\item{alpha}{\code{numeric(1)} in \verb{[0, 1]}. The color transparency.
See also \code{\link[ggplot2:reexports]{ggplot2::alpha()}}.}

\item{color}{\code{character(1)} or \code{integer(1)}. Specify the colors for
lines, points, bar, box, or ROC.}

\item{col}{\code{character(1)} or \code{integer(1)}. Specify the colors for
lines, points, bar, box, or ROC.}

\item{data}{A \code{tibble} or \code{data.frame} object containing data for analysis.
Should often contain a "Response" column indicating the
response variable, often the grouping variable.}

\item{feats}{\code{character(n)}. A vector of features, usually column names
of a data frame.}

\item{formula}{A \code{formula} class object, specifying
the model to be fitted (e.g. \eqn{response ~ x_1 + x_2 + ... + x_n}).}

\item{response}{\code{character(1)}. The column name to use as the
response grouping. Expected to be a factor with 2 levels.
If not a factor, \code{response} will be coerced to a
factor (i.e. alphabetical order). Level 2 of the factor is
considered the \code{disease/case} group, thus positive values
for the statistic indicate up-regulation in \code{Level2} compared \code{Level1}.}

\item{nboot}{\code{integer(1)}. The number of bootstrap estimates to perform.}

\item{r_seed}{`integer(1). The value of the random seed if
reproducibility is desired.}
}
\description{
The parameters below are commonly used throughout
the \pkg{libml} package.
}
