% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Create Manhattan Plot}
\usage{
plot_manhattan(
  data,
  x.lab = "Feature",
  type = c("median", "t.test", "ks.test"),
  as.pvalue = FALSE
)
}
\arguments{
\item{data}{A training data \code{tr_data} object.}

\item{x.lab}{Character. Set the x-axis label.}

\item{type}{Character. The type measure used to evaluate expression change.
Can be one of: "median", "t.test", or "ks.test". Pattern is matched so
abbreviated strings are allowed.}

\item{as.pvalue}{Logical. Should p-values be plotted in linear space
or log10-space?}
}
\description{
Creates a Manhattan plot of the differences or log-ratios
of the training set for each analyte.
}
\examples{
tr <- create_train(sim_test_data,
                   class_response \%in\% c("control", "disease"),
                   group.var = class_response,
                   classes = c("control", "disease"))

# Various options for plotting
plot_manhattan(tr, type = "med")
plot_manhattan(tr, type = "t")
plot_manhattan(tr, type = "ks")
plot_manhattan(tr, type = "t", as.pvalue = TRUE)
}
\seealso{
\code{\link[=calc.ks]{calc.ks()}}, \code{\link[=calc.t]{calc.t()}}
}
\author{
Stu Field
}
