% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-calc-predictions.R, R/s3-get-model-coef.R,
%   R/s3-get-model-features.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{calc_predictions}
\alias{get_model_coef}
\alias{get_model_features}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{globalr}{\code{\link[globalr:s3-generics]{calc_predictions}}, \code{\link[globalr:s3-generics]{get_model_coef}}, \code{\link[globalr:s3-generics]{get_model_features}}}
}}

